### 2025-06-16 #8414 @editorialbot
[PRE REVIEW]: Kleinkram: Open Robotic Data Management
**Submitting author:** @wp99cp (<a href="http://orcid.org/0009-0000-5231-9310">Cyrill Püntener</a>)
**Repository:** https://github.com/leggedrobotics/kleinkram
**Branch with paper.md** (empty if default branch):
**Version:** v0.46.0
