### 2025-07-11 #8414 @boisgera
Hi @wp99cp,

Do you expect the project to be testable on a recent Mac OS with ARM architecture (or maybe on a virtualized Linux for ARM)? The question comes from a potential reviewer.

Best regards,

Sébastien
