### 2025-07-23 #8414 @wp99cp
Version 0.48.0 (currently on the dev branch) offers full compatibility with ARM on macOS. Thanks @johannschwabe for testing and patching.
