### 2025-07-23 2b2f6af
fix #1534
