+++ common/api/types/user.dto.ts
@@ -49,6 +49,17 @@ export class UserDto {
     @IsOptional()
     @IsString()
     avatarUrl!: string | null;
+
+    @ApiProperty({
+        description:
+            'Email address of the user. Set to null if not available or not accessible.',
+        nullable: true,
+        required: false,
+    })
+    @IsNotUndefined()
+    @IsOptional()
+    @IsEmail()
+    email!: string | null;
 }
 
 export class AccessGroupDto {
