### 2025-07-28 70b1b45
add email to search field on exact match
