### 2025-07-28 bddfa9c
improve search by trimming search string
