### 2025-08-04 d027338
fix vit build errors
