### 2025-08-04 084048d
fix seeding for users: correctly add them to their primary group
