### 2025-08-04 81b3d7c
add GitHub as an auth provider
