+++ docs/usage/access-control/base-concepts.md
@@ -39,6 +39,7 @@ assigned for each project. If a user belongs to multiple access groups with vary
 access is applied.
 
 ### Editing Access Groups
+
 The access group can be edited by the creator or any other users with the <CanEdit/> permission on the access group.
 Currently, there is no way to add additional editors to an access group (the creator is always an editor). This must be
 done manually within the database.
