### 2025-08-04 49714c8
add check for correct oauth provider in CLI
