### 2025-08-04 6f33ef2
proper error handling if mail address is linked to a different oAuth provider
