### 2025-08-04 436ae56
fix exception parsing for AuthFlowException
