### 2025-08-04 96e44cc
fix unnecessary type assertion
