+++ backend/src/routing/filters/global-error-filter.ts
@@ -51,8 +51,7 @@ export class GlobalErrorFilter implements ExceptionFilter {
 
         if (exception instanceof AuthFlowException) {
             // If the exception is an AuthFlowException, redirect to the frontend login page
-            const authFlowException: AuthFlowException =
-                exception as AuthFlowException;
+            const authFlowException: AuthFlowException = exception;
 
             logger.debug(
                 `Redirecting to login with error: ${authFlowException.message}`,
