### 2025-08-06 9841fcf
don't check version number for pr to dev
