+++ .github/workflows/check_version_number.yml
@@ -4,7 +4,6 @@ on:
     pull_request:
         branches:
             - main
-            - dev
 
 jobs:
     check-version-number:
