### 2025-08-06 ebbb2a4
improve action triggers
we want to run prettier if the base branch was changed
