+++ .github/workflows/check_version_number.yml
@@ -2,7 +2,8 @@ name: check-version-number
 
 on:
     pull_request:
-        branches:
+      types: [ opened, synchronize, reopened, edited ]
+      branches:
             - main
 
 jobs:
