+++ .github/workflows/eslint.yml
@@ -10,7 +10,7 @@ jobs:
         runs-on: [self-hosted, linux]
 
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v5
 
             - name: Use Node.js
               uses: actions/setup-node@v4
