+++ .github/workflows/draft-pdf.yml
@@ -11,7 +11,7 @@ jobs:
         name: Paper Draft
         steps:
             - name: Checkout
-              uses: actions/checkout@v4
+              uses: actions/checkout@v5
             - name: Build draft PDF
               uses: openjournals/openjournals-draft-action@master
               with:
