+++ .github/workflows/example_actions.yml
@@ -12,7 +12,7 @@ jobs:
         runs-on: ubuntu-latest
 
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v5
             - name: Login to DockerHub
               uses: docker/login-action@v3
               with:
