+++ .github/workflows/test.yml
@@ -17,7 +17,7 @@ jobs:
             pull-requests: write # Allows adding comments to PRs
 
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v5
 
             - name: Use Node.js
               uses: actions/setup-node@v4.4.0
