### 2025-08-14 leggedrobotics/kleinkram#1686 @boisgera
Database Seeding Outcome
JOSS submission context: https://github.com/openjournals/joss-reviews/issues/8636

After having followed the instructions of [Enable Database Seeding] (actually, [seeding is on by default] AFAICT), I was expecting to see something in the datatable section of the dashboard but I do have a "No data available" message.

Is it the expected outcome? Did I miss a step somehow?

[seeding is on by default]: https://github.com/leggedrobotics/kleinkram/blob/main/.env#L26
[Enable Database Seeding]: https://docs.datasets.leggedrobotics.com/development/getting-started.html#enable-database-seeding
