### 2025-08-14 leggedrobotics/kleinkram#1687 @boisgera
Access Control
JOSS Context: https://github.com/openjournals/joss-reviews/issues/8636

Following the instructions in [Access Control], I have changed my `backed/access_config.json` file to:

```json
{
    "emails": [
        {
            "email": "gmail.com",
            "access_groups": ["00000000-0000-0000-0000-000000000000"]
        }
    ],
    "access_groups": [
        {
            "name": "GMail",
            "uuid": "00000000-0000-0000-0000-000000000000",
            "rights": 30
        }
    ]
}
```

and logged into a local instance of Kleinkram using my Google account.

However, I still cannot cannot create a new project or mission, which seems to be necessary to upload new MCAP data.

Am I doing something wrong? Do you need me to provide more info?

[Access Control]: https://docs.datasets.leggedrobotics.com/development/access-control/base-concepts.html
