### 2025-08-14 leggedrobotics/kleinkram#1687 @wp99cp
Yes, it is indeed necessary to first create a project before you can upload any data to Kleinkram. Access permission is then given on a project level. More details can be found in the [docs](https://docs.datasets.leggedrobotics.com/usage/access-control/project.html).

As most of the access rights are managed at the project level, this leads to a special handling for managing permissions to create projects. Thus, we created an additional `rights` field on the group level. This indicated the default and global permission of that group. Level `10` as configured in `access_config.json` means you have global create rights, i.e. you can create new projects and access groups.

By default, this means you must be part of an `affiliation group` to create new projects. Normally, you should be added automatically to an affiliation group based on your affiliation to a particular domain, which is why it's called an affiliation group.

Please verify that you are indeed part of such a group by visiting http://localhost:8003/access-groups, selecting the filter option `Affiliation Groups` and then selecting the group. See the following screenshots:

## How to Resolve Your Issue

Your modifications of the `access_config.json`  file are correct.

Suppose you are still not part of such a group as of the above check. Please consider resetting your database by running `docker compose down --volumes` then starting the local server again. It may be that changing `access_config.json` does not correctly get applied when the application does not start with a fresh/empty database state. I will double-check that and get back to you.

Please let me know if that helps.
