### 2025-08-14 leggedrobotics/kleinkram#1686 @wp99cp
Yes, according to the current state of the `.env` file in this repo, seeding is enabled by default, which is correct.

It may be related to #1687 that you cannot see any of the seeded data, as you may not have the rights to see them. By default, you can only see projects explicitly shared with you or projects part of an affiliation group you are also a member of. If you are not part of any group with associated projects, you won't see any data.

In that case, the message `No data available` is intentional.
