+++ common/seeds/user/create-users.seed.ts
@@ -18,9 +18,9 @@ export default class CreateUsers implements Seeder {
     // Settings for the Seeder
 
     private USER_COUNT = 10;
-    private ACCESS_GROUP_COUNT = 10;
+    private ACCESS_GROUP_COUNT = 20;
     private PROJECT_COUNT = 10;
-    private GROUP_ACCESS_COUNT = 5;
+    private GROUP_ACCESS_COUNT = 2;
     private SEED_ADMINS = false;
     private TAG_TYPE_COUNT = 10;
 
@@ -53,21 +53,30 @@ export default class CreateUsers implements Seeder {
 
         // generate admin users
         // and add them to the affiliation group
-        const adminUsers = await Promise.all(
-            adminMails.map((mail) =>
-                factory(User)({
-                    mail,
-                    role: this.SEED_ADMINS ? UserRole.ADMIN : UserRole.USER,
-                    defaultGroupIds: ['00000000-0000-0000-0000-000000000000'],
-                } as UserContext).create(),
-            ),
-        );
+        const adminUsers =
+            (await Promise.all(
+                adminMails.map((mail) =>
+                    factory(User)({
+                        mail,
+                        role: this.SEED_ADMINS ? UserRole.ADMIN : UserRole.USER,
+                        defaultGroupIds: [
+                            '00000000-0000-0000-0000-000000000000',
+                        ],
+                    } as UserContext).create(),
+                ),
+            )
+                // catch any errors during user creation
+                .catch(console.error)) ?? [];
 
         // generate remaining users
         const remainingUserCount = this.USER_COUNT - adminMails.length;
-        const users = await factory(User)({
-            role: UserRole.USER,
-        }).createMany(remainingUserCount);
+        const users =
+            (await factory(User)({
+                role: UserRole.USER,
+            })
+                .createMany(remainingUserCount)
+                // catch any errors during user creation
+                .catch(console.error)) ?? [];
 
         const allUsers = [...adminUsers, ...users];
 
@@ -79,52 +88,71 @@ export default class CreateUsers implements Seeder {
                     isPersonal: true,
                 } as AccessGroupFactoryContext).create(),
             ),
-        );
+        )
+            // catch any errors during personal access group creation
+            .catch(console.error);
 
         // generate additional groups
-        const groups = await factory(AccessGroup)({
+        const accessGroups = await factory(AccessGroup)({
             allUsers: allUsers,
             isPersonal: false,
         } as AccessGroupFactoryContext).createMany(this.ACCESS_GROUP_COUNT);
 
         // Generate Projects, Missions, Files, and Topics...
-        const projects = await factory(Project)({
-            allAccessGroups: groups,
-            allUsers: allUsers,
-            tagTypes,
-        } as ProjectContext).createMany(this.PROJECT_COUNT);
+        const projects =
+            (await factory(Project)({
+                allAccessGroups: accessGroups,
+                allUsers: allUsers,
+                tagTypes,
+            } as ProjectContext)
+                .createMany(this.PROJECT_COUNT)
+                // catch any errors during project creation
+                .catch(console.error)) ?? [];
 
         // set access rights for projects
         await factory(ProjectAccess)({
             projects: projects,
-            accessGroups: groups,
+            accessGroups: accessGroups,
         }).createMany(this.GROUP_ACCESS_COUNT);
 
         for (const project of projects) {
-            const missions = await factory(Mission)({
-                user: project.creator,
-                project,
-            }).createMany(
-                extendedFaker.number.int({
-                    min: 0,
-                    max: 20,
-                }),
-            );
+            const missions =
+                (await factory(Mission)({
+                    user: project.creator,
+                    project,
+                })
+                    .createMany(
+                        extendedFaker.number.int({
+                            min: 0,
[…]
