### 2025-08-15 leggedrobotics/kleinkram#1687 @wp99cp
That's unfortunate! Project and mission names must be unique in Kleinkram; however, our seeding code generates them randomly. You have hit the case where the seeding script tries to seed two missions with the same name. That is a bug for which I just published a fix to the dev branch; we will release that fix with the next minor version of Kleinkram (0.50.0).

I can also confirm that if the seeding script crashes, the affiliation group isn't created correctly either.

## How to Fix
In the meantime, consider switching to `dev`, applying the patch [19ea76b](https://github.com/leggedrobotics/kleinkram/commit/19ea76b76bfb42c331876b723364f498a72bb458), retrying the seeding crashes probabilistically, or disabling seeding at all.

## Additional Background
Don't expect to see any seeded projects in the dashboard when you log in with your personal Google Account unless you've added your email to the following list. This is intentional: we seed the database before your account was created by signing in to the frontend. Currently, we do not add any seeded projects to the default affiliation group. Thus, it's expected to see an empty dashboard in that case, as there are no projects you have access to.

https://github.com/leggedrobotics/kleinkram/blob/8d7cf5f85e8683d1fd6b1ccf1772af09d65598a5/common/seeds/user/create-users.seed.ts#L52
