### 2025-08-21 c4de3ff
allow to use fake-oauth with cli
