### 2025-09-02 leggedrobotics/kleinkram#1709 @boisgera
Installation instructions variants
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636

[The README instructions][https://github.com/leggedrobotics/kleinkram?tab=readme-ov-file#installation] and the [run locally instruction of the dev docs](https://docs.datasets.leggedrobotics.com/development/getting-started.html#run-the-project-locally) are different.

Could you explain if these differences are meaningful and their consequences?

Should `yarn` be run before `docker compose`? Do I need to add the `--build` option to `docker compose up`?
