### 2025-09-02 leggedrobotics/kleinkram#1710 @boisgera
Failure of docker compose up --build (slow network issue?)
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636

I failed to build the project with `docker compose up` 4 or 5 times in a row with different error logs. The error log of my last attempt is below.

The previous errors suggested that `yarn` was simply giving up  (`There appears to be trouble with your network connection. Retrying...`) which apparently is the kind of messages that yarn outputs [when the network is too slow](https://stackoverflow.com/questions/51508364/yarn-there-appears-to-be-trouble-with-your-network-connection-retrying). The collective wisdown suggests to use the `--network-timeout 100000` with yarn in this case ...

Note that my `yarn` command in the project directory succeeds in 6m20 (with a 15Mbits/s download).

Do you think that the error is related to a (moderately) slow network? Did you evaluate this use case? Any change that I could try easier than changing the network-limit of all yarn calls in the Dockerfile?

```
$  docker compose up --build
[+] Running 1/1
 ! artifactUploader Warning pull access denied for rslethz/grandtour-datasets, repository does not exist or ma...                           1.2s
[+] Building 329.2s (52/68)
 => [internal] load local bake definitions                                                                                                  0.0s
 => => reading from stdin 2.53kB                                                                                                            0.0s
 => [api-server internal] load build definition from Dockerfile                                                                             0.0s
 => => transferring dockerfile: 633B                                                                                                        0.0s
 => [frontend internal] load build definition from dev.Dockerfile                                                                           0.1s
 => => transferring dockerfile: 687B                                                                                                        0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 2)                                                              0.1s
 => [docs internal] load build definition from local-dev.Dockerfile                                                                         0.1s
 => => transferring dockerfile: 166B                                                                                                        0.0s
 => [artifactUploader internal] load build definition from Dockerfile                                                                       0.1s
 => => transferring dockerfile: 636B                                                                                                        0.0s
 => [queue-consumer internal] load build definition from Dockerfile                                                                         0.1s
 => => transferring dockerfile: 703B                                                                                                        0.0s
 => [api-server internal] load metadata for docker.io/library/node:24-alpine                                                                0.8s
 => [artifactUploader internal] load metadata for docker.io/library/python:3.9-slim                                                         1.0s
 => [auth] library/python:pull token for registry-1.docker.io                                                                               0.0s
 => [frontend internal] load .dockerignore                                                                                                  0.0s
 => => transferring context: 82B                                                                                                            0.0s
 => [docs internal] load .dockerignore                                                                                                      0.0s
 => => transferring context: 69B                                                                                                            0.0s
 => [queue-consumer internal] load build context                                                                                            0.1s
 => => transferring context: 556.96kB                                                                                                       0.0s
 => [frontend internal] load build context                                                                                                  0.0s
 => => transferring context: 999.92kB                                                                                                       0.0s
 => [docs  1/15] FROM docker.io/library/node:24-alpine@sha256:be4d5e92ac68483ec71440bf5934865b4b7fcb93588f17a24d411d15f0204e4f              0.0s
 => => resolve docker.io/library/node:24-alpine@sha256:be4d5e92ac68483ec71440bf5934865b4b7fcb93588f17a24d411d15f0204e4f                     0.0s
 => [api-server internal] load build context                                                                                                0.1s
 => => transferring context: 4.01MB                                                                                                         0.1s
 => [docs internal] load build context                                                                                                      0.0s
 => => transferring context: 768B                                                                                                           0.0s
 => CACHED [frontend  2/15] RUN echo "QUASAR_ENDPOINT is set to http://localhost:3000"                                                      0.0s
 => CACHED [frontend  3/15] WORKDIR /app                                                                                                    0.0s
 => CACHED [frontend  4/15] COPY ./common/package.json ./common/                                                                            0.0s
 => CACHED [frontend  5/15] COPY ./common/yarn.lock ./common/                                                                               0.0s
 => CACHED [frontend  6/15] COPY ./frontend/package.json ./frontend/                                                                        0.0s
 => CACHED [frontend  7/15] COPY ./frontend/yarn.lock ./frontend/                                                                           0.0s
 => CACHED [frontend  8/15] COPY ./common/frontend_shared ./common/frontend_shared                                                          0.0s
 => CACHED [frontend  9/15] COPY ./.git/HEAD /app/frontend/.git/HEAD                                                                        0.0s
 => CACHED [frontend 10/15] COPY ./.git/refs/heads/ /app/frontend/.git/refs/heads/                                                          0.0s
 => CACHED [frontend 11/15] WORKDIR /app/common                                                                                             0.0s
 => CACHED [frontend 12/15] RUN yarn --immutable                                                                                            0.0s
 => CACHED [frontend 13/15] WORKDIR /app/frontend                                                                                           0.0s
 => ERROR [frontend 14/15] RUN yarn install --ignore-engines --immutable                                                                  327.7s
 => CACHED [docs 2/4] WORKDIR /app                                                                                                          0.0s
 => CACHED [docs 3/4] COPY package*.json ./                                                                                                 0.0s
 => [docs 4/4] RUN yarn install                                                                                                           274.3s
 => [artifactUploader internal] load .dockerignore                                                                                          0.1s
 => => transferring context: 2B                                                                                                             0.0s
 => CACHED [queue-consumer  2/13] COPY ./queueConsumer/package.json /usr/src/queueConsumer/                                                 0.0s
 => CACHED [queue-consumer  3/13] COPY ./queueConsumer/yarn.lock /usr/src/queueConsumer/                                                    0.0s
 => CACHED [queue-consumer  4/13] COPY ./common/package.json /usr/src/common/                                                               0.0s
 => CACHED [queue-consumer  5/13] COPY ./common/yarn.lock /usr/src/common/                                                                  0.0s
 => CACHED [queue-consumer  6/13] WORKDIR /usr/src/queueConsumer                                                                            0.0s
 => CACHED [queue-consumer  7/13] RUN yarn --immutable                                                                                      0.0s
 => CANCELED [queue-consumer  8/13] RUN wget https://github.com/foxglove/mcap/releases/download/releases%2Fmcap-cli%2Fv0.0.42/mcap-linux  327.8s
 => CACHED [api-server  2/14] COPY ./backend/package.json /usr/src/app/backend/                                                             0.0s
 => CACHED [api-server  3/14] COPY ./backend/yarn.lock /usr/src/app/backend/                                                                0.0s
 => CACHED [api-server  4/14] COPY ./common/package.json /usr/src/app/common/                                                               0.0s
 => CACHED [api-server  5/14] COPY ./common/yarn.lock /usr/src/app/common/                                                                  0.0s
 => CACHED [api-server  6/14] WORKDIR /usr/src/app/backend                                                                                  0.0s
 => CANCELED [api-server  7/14] RUN yarn --immutable                                                                                      327.8s
 => [artifactUploader 1/6] FROM docker.io/library/python:3.9-slim@sha256:914169c7c8398b1b90c0b0ff921c8027445e39d7c25dc440337e56ce0f2566e6   0.2s
 => => resolve docker.io/library/python:3.9-slim@sha256:914169c7c8398b1b90c0b0ff921c8027445e39d7c25dc440337e56ce0f2566e6                    0.1s
 => [artifactUploader internal] load build context                                                                                          0.1s
 => => transferring context: 4.00kB                                                                                                         0.0s
 => CACHED [artifactUploader 2/6] WORKDIR /app                                                                                              0.0s
 => CANCELED [artifactUploader 3/6] RUN apt-get update && apt-get install -y     curl     unzip     && rm -rf /var/lib/apt/lists/*        327.6s
 => [docs] exporting to image                                                                                                              27.0s
 => => exporting layers                                                                                                                    22.5s
 => => exporting manifest sha256:6b80691503c64bb906886b33004030594b7dbc032fd612cf00f0eeaa7e238d25                                           0.0s
 => => exporting config sha256:0817411b93b551fc926007b474fd3a59585cb92f26ce5fae060279f18705f858                                             0.0s
 => => exporting attestation manifest sha256:8e1d69d62600c6bc6efa529d2c721d0361ae986364ff4b30ab4390d17d0c8491                               0.0s
 => => exporting manifest list sha256:d74adb7007387cc887e71e7723bf760b4bc313b8320fa82c29931bfd7ef4b5b9                                      0.0s
 => => naming to docker.io/library/kleinkram-docs:latest                                                                                    0.0s
 => => unpacking to docker.io/library/kleinkram-docs:latest                                                                                 4.3s
 => [docs] resolving provenance for metadata file                                                                                           0.0s
------
 > [frontend 14/15] RUN yarn install --ignore-engines --immutable:
0.594 yarn install v1.22.22
0.636 [1/4] Resolving packages...
0.752 (node:1) [DEP0169] DeprecationWarning: `url.parse()` behavior is not standardized and prone to errors that have security implications. Use the WHATWG URL API instead. CVEs are not issued for `url.parse()` vulnerabilities.
0.752 (Use `node --trace-deprecation ...` to show where the warning was created)
0.753 [2/4] Fetching packages...
6.177 error Error: getaddrinfo EAI_AGAIN registry.yarnpkg.com
6.177     at GetAddrInfoReqWrap.onlookupall [as oncomplete] (node:dns:122:26)
6.177 info Visit https://yarnpkg.com/en/docs/cli/install for documentation about this command.
------
dev.Dockerfile:26

--------------------

  24 |

  25 |     WORKDIR /app/frontend

  26 | >>> RUN yarn install --ignore-engines --immutable

  27 |

  28 |     COPY ./frontend/. ./

--------------------

target frontend: failed to solve: process "/bin/sh -c yarn install --ignore-engines --immutable" did not complete successfully: exit code: 1
```
