+++ dev-oauth/fake_oauth.py
@@ -9,48 +9,53 @@ FAKE_USERS = [
         "email": "admin@kleinkram.leggedrobotics.com",
         "displayName": "Kleinkram Admin User Nr. 1",
         "photo": "https://randomuser.me/api/portraits/men/8.jpg",
-        "comment": "This user has admin access to Kleinkram. It sees all seeded projects."
+        "comment": "This user has admin access to Kleinkram. It sees all seeded projects.",
     },
     {
         "id": "2",
         "email": "internal-user@kleinkram.leggedrobotics.com",
         "displayName": "Kleinkram User Nr. 2",
         "photo": "https://randomuser.me/api/portraits/women/71.jpg",
-        "comment": "This user is an internal user of Kleinkram. She is part of an affiliation group and can create projects. This user sees no seeded projects."
+        "comment": "This user is an internal user of Kleinkram. She is part of an affiliation group and can create projects. This user sees no seeded projects.",
     },
     {
         "id": "3",
         "email": "external-user@example.com",
         "displayName": "External Kleinkram User Nr. 3",
         "photo": "https://randomuser.me/api/portraits/women/88.jpg",
-        "comment": "This user is an external user of Kleinkram. She cannot create any new projects, be default she sees no projects."
+        "comment": "This user is an external user of Kleinkram. She cannot create any new projects, be default she sees no projects.",
     },
 ]
 
 # Store issued tokens
 ACTIVE_TOKENS = {}
 
+
 @app.route("/")
 def home():
     """Homepage with a fake login screen"""
     return render_template("login.html", users=FAKE_USERS)
+
 
 @app.route("/oauth/authorize")
 def authorize():
     """OAuth2 Authorization Endpoint - User selects an account"""
     client_id = request.args.get("client_id")
     redirect_uri = request.args.get("redirect_uri")
-    state = request.args.get("state") # Get the state parameter
+    state = request.args.get("state")  # Get the state parameter
 
     # Pass the state parameter to the template
-    return render_template("login.html", users=FAKE_USERS, redirect_uri=redirect_uri, state=state)
+    return render_template(
+        "login.html", users=FAKE_USERS, redirect_uri=redirect_uri, state=state
+    )
+
 
 @app.route("/login", methods=["POST"])
 def login():
     """Handles login and redirects to the given redirect_uri with a fake code"""
     user_id = request.form.get("user_id")
     redirect_uri = request.form.get("redirect_uri")
-    state = request.form.get("state") # Get the state parameter from the form
+    state = request.form.get("state")  # Get the state parameter from the form
 
     user = next((u for u in FAKE_USERS if u["id"] == user_id), None)
     if not user:
@@ -65,6 +70,7 @@ def login():
         redirect_url += f"&state={state}"
 
     return redirect(redirect_url)
+
 
 @app.route("/oauth/token", methods=["POST"])
 def token():
@@ -93,6 +99,7 @@ def token():
     }
     return jsonify(response)
 
+
 @app.route("/oauth/profile", methods=["GET"])
 def profile():
     """Returns the user profile based on the access token"""
@@ -107,6 +114,7 @@ def profile():
         return jsonify({"error": "Invalid or expired token"}), 401
 
     return jsonify(user)
+
 
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=5000, debug=True)
