### 2025-09-04 leggedrobotics/kleinkram#1709 @wp99cp
I have unified the instructions in [4dfec62](https://github.com/leggedrobotics/kleinkram/commit/4dfec623e42d621c1c1024f86dcd2f298e3d42da).

**The difference is the following:**

`docker compose up` and `docker compose up --build` will start the application, where the latter forces Docker Compose to rebuild the images based on the local changes/code base. Without `--build`, we may pull the images from DockerHub instead.

However, in practice, this doesn't make a big difference, as the images published on DockerHub are private anyway, thus `docker compose up` falls back to building the images from source.

The --watch flag is optional and is used to watch for changes in the codebase and enables [live and hot-reloading](https://stackoverflow.com/a/41429055). For both the backend and frontend. This is used for local development.

**When to run yarn install**

As stated in the [docs](https://docs.datasets.dev.leggedrobotics.com/development/getting-started.html#setting-up-the-development-environment), you need to run `yarn install` outside of the docker container in order to get code completion and linting in your IDE. If you want to run the application locally, you don't need to run `yarn install`; this is only required for local development.
