### 2025-09-08 leggedrobotics/kleinkram#1710 @boisgera
Same kind of error using Cloudflare DNS on the host machine (but much faster). Could it be a Docker network configuration issue on my computer?

```
------
 > [frontend 12/15] RUN yarn --immutable:
0.754 yarn install v1.22.22
0.830 [1/4] Resolving packages...
0.891 (node:1) [DEP0169] DeprecationWarning: `url.parse()` behavior is not standardized and prone to errors that have security implications. Use the WHATWG URL API instead. CVEs are not issued for `url.parse()` vulnerabilities.
0.891 (Use `node --trace-deprecation ...` to show where the warning was created)
0.892 [2/4] Fetching packages...
6.106 error Error: getaddrinfo EAI_AGAIN registry.yarnpkg.com
6.106     at GetAddrInfoReqWrap.onlookupall [as oncomplete] (node:dns:122:26)
6.106 info Visit https://yarnpkg.com/en/docs/cli/install for documentation about this command.
------
```
