### 2025-09-08 leggedrobotics/kleinkram#1710 @boisgera
> Can you run `yarn install`outside of the Docker container, i.e. directly in the root directory of the project?

It works but I already have the `info There appears to be trouble with your network connection. Retrying...` message.

**Update.** Did it again (from a fresh kleinkram repo, to be sure). This time no network warning message.

```
💻  yarn install
yarn install v1.22.22
[1/4] Resolving packages...
[2/4] Fetching packages...
warning np@7.7.0: The engine "git" appears to be invalid.
[3/4] Linking dependencies...
[4/4] Building fresh packages...
$ (cd backend && yarn install --ignore-engines ); (cd frontend && yarn install --ignore-engines ); (cd queueConsumer && yarn install --ignore-engines ); (cd common && yarn install --ignore-engines ); (cd docs && yarn install --ignore-engines )
yarn install v1.22.22
[1/4] Resolving packages...
[2/4] Fetching packages...
info There appears to be trouble with your network connection. Retrying...
[3/4] Linking dependencies...
[4/4] Building fresh packages...
Done in 87.92s.
yarn install v1.22.22
[1/4] Resolving packages...
[2/4] Fetching packages...
[3/4] Linking dependencies...
[4/4] Building fresh packages...
Done in 5.99s.
yarn install v1.22.22
[1/4] Resolving packages...
[2/4] Fetching packages...
[3/4] Linking dependencies...
[4/4] Building fresh packages...
Done in 20.25s.
yarn install v1.22.22
[1/4] Resolving packages...
[2/4] Fetching packages...
[3/4] Linking dependencies...
warning " > @nestjs/swagger@11.2.0" has unmet peer dependency "@nestjs/common@^11.0.1".
warning " > @nestjs/swagger@11.2.0" has unmet peer dependency "@nestjs/core@^11.0.1".
warning " > @nestjs/swagger@11.2.0" has unmet peer dependency "reflect-metadata@^0.1.12 || ^0.2.0".
warning "@nestjs/swagger > @nestjs/mapped-types@2.1.0" has unmet peer dependency "@nestjs/common@^10.0.0 || ^11.0.0".
warning "@nestjs/swagger > @nestjs/mapped-types@2.1.0" has unmet peer dependency "reflect-metadata@^0.1.12 || ^0.2.0".
warning " > typeorm-seeding@1.6.1" has incorrect peer dependency "typeorm@^0.2.24".
[4/4] Building fresh packages...
Done in 3.32s.
yarn install v1.22.22
[1/4] Resolving packages...
[2/4] Fetching packages...
[3/4] Linking dependencies...
warning "vitepress > @docsearch/js > @docsearch/react > @algolia/autocomplete-preset-algolia@1.17.7" has unmet peer dependency "@algolia/client-search@>= 4.9.1 < 6".
warning "vitepress > @docsearch/js > @docsearch/react > @algolia/autocomplete-core > @algolia/autocomplete-plugin-algolia-insights@1.17.7" has unmet peer dependency "search-insights@>= 1 < 3".
warning "vitepress > @docsearch/js > @docsearch/react > @algolia/autocomplete-core > @algolia/autocomplete-shared@1.17.7" has unmet peer dependency "@algolia/client-search@>= 4.9.1 < 6".
[4/4] Building fresh packages...
Done in 9.19s.
Done in 130.05s.
```
