### 2025-09-09 8bbb693
feat: display pbar update during downloads
