### 2025-09-10 leggedrobotics/kleinkram#1710 @boisgera
**Update.** I am having a similar (yarn registry?) issue with the same computer on a fast/reliable network.

However this time I managed to build+run at my second attempt (while it was impossible before).

Full log:

```
docker compose up --build
[+] Building 36.0s (70/79)
 => [internal] load local bake definitions                                                                                                  0.0s
 => => reading from stdin 2.99kB                                                                                                            0.0s
 => [frontend internal] load build definition from dev.Dockerfile                                                                           0.1s
 => => transferring dockerfile: 687B                                                                                                        0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 2)                                                              0.1s
 => [queue-consumer internal] load build definition from Dockerfile                                                                         0.1s
 => => transferring dockerfile: 703B                                                                                                        0.0s
 => [artifactUploader internal] load build definition from Dockerfile                                                                       0.1s
 => => transferring dockerfile: 636B                                                                                                        0.0s
 => [fake-oauth internal] load build definition from Dockerfile                                                                             0.1s
 => => transferring dockerfile: 367B                                                                                                        0.0s
 => [docs internal] load build definition from local-dev.Dockerfile                                                                         0.1s
 => => transferring dockerfile: 166B                                                                                                        0.0s
 => [api-server internal] load build definition from Dockerfile                                                                             0.1s
 => => transferring dockerfile: 633B                                                                                                        0.0s
 => [api-server internal] load metadata for docker.io/library/node:24-alpine                                                                1.2s
 => [artifactUploader internal] load metadata for docker.io/library/python:3.9-slim                                                         1.9s
 => [fake-oauth internal] load metadata for docker.io/library/python:3.10-slim                                                              2.1s
 => [auth] library/node:pull token for registry-1.docker.io                                                                                 0.0s
 => [auth] library/python:pull token for registry-1.docker.io                                                                               0.0s
 => [docs internal] load .dockerignore                                                                                                      0.0s
 => => transferring context: 69B                                                                                                            0.0s
 => [frontend internal] load .dockerignore                                                                                                  0.0s
 => => transferring context: 82B                                                                                                            0.0s
 => [queue-consumer 1/4] FROM docker.io/library/node:24-alpine@sha256:be4d5e92ac68483ec71440bf5934865b4b7fcb93588f17a24d411d15f0204e4f      0.0s
 => => resolve docker.io/library/node:24-alpine@sha256:be4d5e92ac68483ec71440bf5934865b4b7fcb93588f17a24d411d15f0204e4f                     0.0s
 => [docs internal] load build context                                                                                                      0.0s
 => => transferring context: 769B                                                                                                           0.0s
 => [frontend internal] load build context                                                                                                  0.0s
 => => transferring context: 1.00MB                                                                                                         0.0s
 => [queue-consumer internal] load build context                                                                                            0.1s
 => => transferring context: 558.98kB                                                                                                       0.0s
 => [api-server internal] load build context                                                                                                0.1s
 => => transferring context: 4.01MB                                                                                                         0.1s
 => CACHED [docs 2/4] WORKDIR /app                                                                                                          0.0s
 => CACHED [docs 3/4] COPY package*.json ./                                                                                                 0.0s
 => CACHED [frontend  2/15] RUN echo "QUASAR_ENDPOINT is set to http://localhost:3000"                                                      0.0s
 => CACHED [frontend  3/15] WORKDIR /app                                                                                                    0.0s
 => CACHED [frontend  4/15] COPY ./common/package.json ./common/                                                                            0.0s
 => CACHED [frontend  5/15] COPY ./common/yarn.lock ./common/                                                                               0.0s
 => CACHED [frontend  6/15] COPY ./frontend/package.json ./frontend/                                                                        0.0s
 => CACHED [frontend  7/15] COPY ./frontend/yarn.lock ./frontend/                                                                           0.0s
 => CACHED [frontend  8/15] COPY ./common/frontend_shared ./common/frontend_shared                                                          0.0s
 => CACHED [frontend  9/15] COPY ./.git/HEAD /app/frontend/.git/HEAD                                                                        0.0s
 => CACHED [frontend 10/15] COPY ./.git/refs/heads/ /app/frontend/.git/refs/heads/                                                          0.0s
 => CACHED [frontend 11/15] WORKDIR /app/common                                                                                             0.0s
 => CACHED [frontend 12/15] RUN yarn --immutable                                                                                            0.0s
 => CACHED [frontend 13/15] WORKDIR /app/frontend                                                                                           0.0s
 => ERROR [frontend 14/15] RUN yarn install --ignore-engines --immutable                                                                   34.0s
 => [docs 4/4] RUN yarn install                                                                                                            25.2s
 => CACHED [queue-consumer  2/13] COPY ./queueConsumer/package.json /usr/src/queueConsumer/                                                 0.0s
 => CACHED [queue-consumer  3/13] COPY ./queueConsumer/yarn.lock /usr/src/queueConsumer/                                                    0.0s
 => CACHED [queue-consumer  4/13] COPY ./common/package.json /usr/src/common/                                                               0.0s
 => CACHED [queue-consumer  5/13] COPY ./common/yarn.lock /usr/src/common/                                                                  0.0s
 => CACHED [queue-consumer  6/13] WORKDIR /usr/src/queueConsumer                                                                            0.0s
 => CACHED [queue-consumer  7/13] RUN yarn --immutable                                                                                      0.0s
 => [queue-consumer  8/13] RUN wget https://github.com/foxglove/mcap/releases/download/releases%2Fmcap-cli%2Fv0.0.42/mcap-linux-amd64 -O /  4.5s
 => CACHED [api-server  2/14] COPY ./backend/package.json /usr/src/app/backend/                                                             0.0s
 => CACHED [api-server  3/14] COPY ./backend/yarn.lock /usr/src/app/backend/                                                                0.0s
 => CACHED [api-server  4/14] COPY ./common/package.json /usr/src/app/common/                                                               0.0s
 => CACHED [api-server  5/14] COPY ./common/yarn.lock /usr/src/app/common/                                                                  0.0s
 => CACHED [api-server  6/14] WORKDIR /usr/src/app/backend                                                                                  0.0s
 => CANCELED [api-server  7/14] RUN yarn --immutable                                                                                       34.3s
 => [artifactUploader internal] load .dockerignore                                                                                          0.0s
 => => transferring context: 2B                                                                                                             0.0s
 => [artifactUploader 1/6] FROM docker.io/library/python:3.9-slim@sha256:71b97a8eed69cddd132431327ff7c204fd6ea3d38e0c5d11d52f9661482bc8a3   4.0s
 => => resolve docker.io/library/python:3.9-slim@sha256:71b97a8eed69cddd132431327ff7c204fd6ea3d38e0c5d11d52f9661482bc8a3                    0.0s
 => => sha256:7fcdf9369fa96e0413fe19da3d316fb6c3bfb0d7371fa4ce617617cac3e8de12 249B / 249B                                                  0.6s
 => => sha256:41dc2499d8fe1ea2351cc01f3716ce6a95ad0e9bf90c0819fd0c4a93cf4e9b24 13.37MB / 13.37MB                                            2.4s
 => => sha256:1d454ace0e384876850a0aa5ef6b8c45705445114ab233959bdab71a577b9200 1.29MB / 1.29MB                                              0.9s
 => => sha256:ce1261c6d567efa8e3b457673eeeb474a0a8066df6bb95ca9a6a94a31e219dd3 29.77MB / 29.77MB                                            2.6s
 => => extracting sha256:ce1261c6d567efa8e3b457673eeeb474a0a8066df6bb95ca9a6a94a31e219dd3                                                   0.5s
 => => extracting sha256:1d454ace0e384876850a0aa5ef6b8c45705445114ab233959bdab71a577b9200                                                   0.1s
 => => extracting sha256:41dc2499d8fe1ea2351cc01f3716ce6a95ad0e9bf90c0819fd0c4a93cf4e9b24                                                   0.3s
 => => extracting sha256:7fcdf9369fa96e0413fe19da3d316fb6c3bfb0d7371fa4ce617617cac3e8de12                                                   0.2s
 => [artifactUploader internal] load build context                                                                                          0.0s
 => => transferring context: 4.00kB                                                                                                         0.0s
 => [fake-oauth internal] load .dockerignore                                                                                                0.0s
 => => transferring context: 2B                                                                                                             0.0s
 => [fake-oauth 1/4] FROM docker.io/library/python:3.10-slim@sha256:122c1a0e792fad67b870205fd0f5e4d6d0f6f3f13b2fce1b9472c1ecbe274671        3.2s
 => => resolve docker.io/library/python:3.10-slim@sha256:122c1a0e792fad67b870205fd0f5e4d6d0f6f3f13b2fce1b9472c1ecbe274671                   0.1s
 => => sha256:fceec8b8a90ef53ff8f6211082b139436e3b3f0f16adf7752fe0d9e6681a3231 250B / 250B                                                  0.3s
 => => sha256:6dc114d5e12c107652ba61ef068cfacc611596390cbafa6cd21206a670e4e264 14.09MB / 14.09MB                                            1.7s
 => => sha256:d02c566428bb557aab93bf3d9d768b9abb145010f1f8a6756db930f0902f60f7 1.29MB / 1.29MB                                              0.4s
 => => extracting sha256:d02c566428bb557aab93bf3d9d768b9abb145010f1f8a6756db930f0902f60f7                                                   0.1s
 => => extracting sha256:6dc114d5e12c107652ba61ef068cfacc611596390cbafa6cd21206a670e4e264                                                   0.4s
 => => extracting sha256:fceec8b8a90ef53ff8f6211082b139436e3b3f0f16adf7752fe0d9e6681a3231                                                   0.1s
 => [fake-oauth internal] load build context                                                                                                0.0s
 => => transferring context: 6.16kB                                                                                                         0.0s
 => [queue-consumer  9/13] WORKDIR /usr/src/common                                                                                          0.2s
 => [artifactUploader 2/6] WORKDIR /app                                                                                                     0.1s
 => [fake-oauth 2/4] WORKDIR /app                                                                                                           0.1s
 => [artifactUploader 3/6] RUN apt-get update && apt-get install -y     curl     unzip     && rm -rf /var/lib/apt/lists/*                  19.6s
 => [fake-oauth 3/4] RUN pip install flask                                                                                                  4.4s
 => [queue-consumer 10/13] RUN yarn --immutable                                                                                             9.2s
 => [fake-oauth 4/4] COPY . .                                                                                                               0.2s
 => [fake-oauth] exporting to image                                                                                                         1.2s
 => => exporting layers                                                                                                                     0.7s
 => => exporting manifest sha256:336c47a5e0b363810d5998d99004a324babc34fcb6aea880e1eb3bb1de95fe65                                           0.0s
 => => exporting config sha256:2b5935d6d49df5e8258feedf36e0e8c5e739bea482050b4b31c5bcd8a38062c9                                             0.0s
 => => exporting attestation manifest sha256:0ea8dc9fed31d49574bb2f0219d5ce25368a671c353e13037520933dc3f59ae1                               0.1s
 => => exporting manifest list sha256:e892edf9ecbd02a060f71c657ac098a09dd50fb8ba13392df4d59a391d629925                                      0.0s
 => => naming to docker.io/library/kleinkram-fake-oauth:latest                                                                              0.0s
 => => unpacking to docker.io/library/kleinkram-fake-oauth:latest                                                                           0.3s
 => [fake-oauth] resolving provenance for metadata file                                                                                     0.0s
 => [queue-consumer 11/13] COPY ./queueConsumer /usr/src/queueConsumer                                                                      0.3s
 => [queue-consumer 12/13] COPY ./common /usr/src/common                                                                                    0.1s
 => [queue-consumer 13/13] WORKDIR /usr/src/queueConsumer                                                                                   0.1s
 => CANCELED [queue-consumer] exporting to image                                                                                           19.7s
 => => exporting layers                                                                                                                    19.7s
 => [artifactUploader 4/6] COPY requirements.txt .                                                                                          0.2s
 => CANCELED [artifactUploader 5/6] RUN pip install --no-cache-dir -r requirements.txt                                                      9.8s
 => CANCELED [docs] exporting to image                                                                                                      8.9s
 => => exporting layers                                                                                                                     8.9s
------
 > [frontend 14/15] RUN yarn install --ignore-engines --immutable:
0.624 yarn install v1.22.22
0.672 [1/4] Resolving packages...
0.796 (node:1) [DEP0169] DeprecationWarning: `url.parse()` behavior is not standardized and prone to errors that have security implications. Use the WHATWG URL API instead. CVEs are not issued for `url.parse()` vulnerabilities.
0.796 (Use `node --trace-deprecation ...` to show where the warning was created)
0.797 [2/4] Fetching packages...
6.002 error Error: getaddrinfo EAI_AGAIN registry.yarnpkg.com
6.002     at GetAddrInfoReqWrap.onlookupall [as oncomplete] (node:dns:122:26)
6.002 info Visit https://yarnpkg.com/en/docs/cli/install for documentation about this command.
------
dev.Dockerfile:26

--------------------

  24 |

  25 |     WORKDIR /app/frontend

  26 | >>> RUN yarn install --ignore-engines --immutable

  27 |

  28 |     COPY ./frontend/. ./

--------------------

target frontend: failed to solve: process "/bin/sh -c yarn install --ignore-engines --immutable" did not complete successfully: exit code: 1

View build details: docker-desktop://dashboard/build/desktop-linux/desktop-linux/whuvwhl9j1c8mwv4st8ba0ew0
```
