### 2025-09-11 leggedrobotics/kleinkram#1733 @boisgera
Unresponsive dashboard
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636

The dashboard web page is frequently unresponsive right after I sign it; that happened to me with Firefox and Chromium.

Reloading the page seems to fix it (?).

I restarted everything and have waited until the log was stable to start the browser but it still happens (and the log doesn't display any extra information that could explain the issue)

Have you already seen this behavior?
