### 2025-09-12 leggedrobotics/kleinkram#1734 @c-joly
Impossible to login with GitHub
This is an issue related to JOSS review https://github.com/openjournals/joss-reviews/issues/8636

I tried to access to the dashboard by login with GitHub. However, I received a message saying that `redirect_uri` is not associated with the application. Is there something specific I should do prior to connect with my GitHub account?
