### 2025-09-12 leggedrobotics/kleinkram#1735 @c-joly
Impossible to login with "Dev Login"
Hi, this is an issue related to JOSS review https://github.com/openjournals/joss-reviews/issues/8636

After I failed to login with GitHub, I tried to login to the dashboard with the Dev Login but each time I click on one of the three proposed users, I come back to the main login page (the one that proposes Google, GitHub and Dev...). I know that @boisgera succeeded to enter the dashboard with that way... FYI, I am using MacOS and the building process with docker did not show any problem.

Here are the log that appears during my try. The first "fake-oath-1" log comes just after I clicked on "Dev Login", and the "login successful" log appear just after I clicked on the user.

```console
kleinkram-api-server         | [DEBUG]: Check Client Version for call to endpoint: /user/me is: 0.50.0
fake-oauth-1                 | 185.199.110.133 - - [12/Sep/2025 12:29:27] "GET /oauth/authorize?response_type=code&redirect_uri=http://localhost:3000/auth/fake-oauth/callback&scope=&client_id=some-random-string-it-does-not-matter HTTP/1.1" 200 -
fake-oauth-1                 | 185.199.110.133 - - [12/Sep/2025 12:29:29] "POST /login HTTP/1.1" 302 -
fake-oauth-1                 | 172.18.0.12 - - [12/Sep/2025 12:29:29] "POST /oauth/token HTTP/1.1" 200 -
fake-oauth-1                 | 172.18.0.12 - - [12/Sep/2025 12:29:29] "GET /oauth/profile HTTP/1.1" 200 -
kleinkram-api-server         | [DEBUG]: Login successful for 4c1b86ec-0d50-42fd-ac63-dc9d52ce6151
kleinkram-api-server         | [DEBUG]: Check Client Version for call to endpoint: /user/me is: 0.50.0
kleinkram-api-server         | [DEBUG]: Check Client Version for call to endpoint: /user/me is: 0.50.0
kleinkram-api-server         | [DEBUG]: Check Client Version for call to endpoint: /user/me is: 0.50.0
kleinkram-queue-consumer     | [DEBUG]: Cleanup containers and dangling actions...
```
