### 2025-09-12 leggedrobotics/kleinkram#1736 @c-joly
API webpage not working
Hi, this is an issue related to JOSS review https://github.com/openjournals/joss-reviews/issues/8636

When I try to connect to `localhost:3000` which is supposed to be the API page, I have a webpage that only contains the following line...

```
{"statusCode":426,"message":"Client version undefined is not a valid version."}
```
