### 2025-09-12 7951eec
feat: adapt mission update command to receive metadata json without specifying Tagtype
