### 2025-09-12 75a1bab
feat: update mission info to display the correct info with metadata table
