### 2025-09-15 286429a
address #1734
