+++ .github/workflows/check_version_number.yml
@@ -2,8 +2,8 @@ name: check-version-number
 
 on:
     pull_request:
-      types: [ opened, synchronize, reopened, edited ]
-      branches:
+        types: [opened, synchronize, reopened, edited]
+        branches:
             - main
 
 jobs:
@@ -18,6 +18,7 @@ jobs:
             - uses: actions/setup-node@v5
               with:
                   node-version: '22'
+                  package-manager-cache: false
 
             - name: Validate Version Number by calling grunt validateVersions
               run: |
