+++ .github/workflows/create_release.yml
@@ -18,6 +18,7 @@ jobs:
             - uses: actions/setup-node@v5
               with:
                   node-version: '22'
+                  package-manager-cache: false
 
             - name: Load Version of Latest Release
               id: check # This will be the reference for getting the outputs.
