### 2025-09-15 leggedrobotics/kleinkram#1739 @c-joly
Creating a mission with the top right button does not work
JOSS submission context: https://github.com/openjournals/joss-reviews/issues/8636

I have a problem to create a mission when I use the blue button "New" in the top right of the dashboard. When I click on "Mission", the Window that appears does not have any editable button (e.g. impossible to chose the project), so that it is impossible to proceed (first screenshot). However, using the button "Create mission" which is inside a project works well (second screenshot).
