### 2025-09-16 leggedrobotics/kleinkram#1757 @boisgera
Hosting actions in arbitrary Docker Hub namespaces
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636
cc: @c-joly

If I understand correctly, all customs actions must be (built locally then) pushed to the rslethz namespace on Dockerhub to be usable in kleinkram. Does that mean that despite the fact that anynone can host their own kleinkram server you need to grant access to each developer that wishes to develop its actions?

Is there a technical reasons for this restriction? Could you lift it?
