### 2025-09-16 leggedrobotics/kleinkram#1757 @c-joly
Hi, I tried to follow the procedure on https://docs.datasets.leggedrobotics.com/usage/actions/write-custom-actions.html to build and publish a basic custom action.
The pushing stage failed with:
``push access denied, repository does not exist or may require authorization: server message: insufficient_scope: authorization failed``
