### 2025-09-17 #8636 @wp99cp
> Hi [@wp99cp](https://github.com/wp99cp)!
>
> I am looking into the author list of your paper. For yourself, Johann Schwabe an Dominique Garnier, the contribution to the project are obvious (from https://github.com/leggedrobotics/kleinkram/graphs/contributors).
>
> * Could you tell me more about the contributions of Jonas Frey and Marco Hutter? (non-code contributions are fine, see [here](https://joss.readthedocs.io/en/latest/submitting.html#authorship) for the details).
> * Did you consider putting Marvin Lichtsteiner in the author list? Was he contacted? His code contribution appears to be significant (at first sight).

Both Jonas Frey and Marco Hutter had contributed to the project through guidance and technical feedback. Notably, Jonas Frey made significant contributions by designing the higher-level features and workflows of Kleinkram. Yes, we did consider adding Marvin Lichtsteiner to the project. Still, we decided against that as he only contributed minor fixes and did not actively contribute towards the core development of Kleinkram.
