### 2025-09-17 57d7858
feat: add check to ensure update metadata types are coherent
