### 2025-09-19 leggedrobotics/kleinkram#1759 @boisgera
Actions Documentation
JOSS Context: https://github.com/openjournals/joss-reviews/issues/8636
cc: @c-joly

Some (subjective) feedback about using the documentation about actions. I had the feeling that there is a subtle disconnect between how the documentation is structured now and how it would have made sense to me when being introduced to actions.

A better user experience would be IMHO:
  - do not use "Write custom actions?" as a title in the front page but something like "Use actions" (or "Use and write actions", anything that makes writing custom actions a 2nd step to smooth the learning curve)
  - in the Actions "Getting Started", remove (or delay/de-emphasize) the mentions of the technologies underlying actions (e.g. "Docker", impact on the container memory, etc). Develop a few (2-5?) generic actions that are already available to every kleinkram user (among the tasks "validate data integrity, extract sensor metadata, generate preview visualisations, convert formats, run benchmarking script" that are mentioned in the paper) and describe the step-by-step procedure that the user needs to go through to use these actions (a screenshot of the interface could be really nice here!)
  - Go through the more technical aspects either in the "Write custom actions" section or in an 3rd intermediate "under the hood" section.
  - In "Write custom actions", show how the built-in actions demonstrated in the "Getting started" section have been defined.

I do *not* suggest that you *have to* refactor your documentation to follow *exactly* this documentation refactoring to get your submission accepted in JOSS. But there are probably among my suggestions some low-hanging fruits that resonate with your vision of the project and that you could implement without to much effort.

In the context of JOSS however, we have to be able to check the functional claims of the paper and therefore at this stage I would really appreciate a few complete examples of actions that perform some of the tasks mentioned in the paper (e.g. "convert formats", "extract sensor metadata", etc.) and as I already mentioned they could be really valuable to the first-time user.

A final idea/remark (not a requirement for JOSS acceptance!): I have the feeling that many potential kleinkram users will not have the devops/admin culture that provides familiarity with Docker, continuous integretion, etc. How could the project be easier to get into for a person that knows "only" about Python and its mcap library for example (with *maybe* some of pip, requirements.txt, pyproject.tom and/or uv for the tooling?)? Would a generic Python template action make sense to bridge this gap for these users?

Best regards,

Sébastien
