### 2025-09-23 56bda59
feat: project get returns required tags too
