+++ common/api/types/project/project-with-missions.dto.ts
@@ -2,9 +2,9 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
 import { FlatMissionDto } from '../mission/mission.dto';
-import { ProjectWithRequiredTags } from './project-with-required-tags';
+import { ProjectWithRequiredTagsDto } from './project-with-required-tags.dto';
 
-export class ProjectWithMissionsDto extends ProjectWithRequiredTags {
+export class ProjectWithMissionsDto extends ProjectWithRequiredTagsDto {
     @ApiProperty({
         type: () => [FlatMissionDto],
         description: 'List of missions',
