### 2025-09-23 89f779d
feat: create mission validates metadata against required metadata
