### 2025-09-24 0c2377c
chore: validate mission after creation
