+++ cli/kleinkram/api/routes.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
 import json
+import tempfile
 from enum import Enum
+from pathlib import Path
 from typing import Any
 from typing import Dict
 from typing import Generator
@@ -40,10 +42,11 @@ from kleinkram.errors import MissionExists
 from kleinkram.errors import MissionNotFound
 from kleinkram.errors import ProjectExists
 from kleinkram.errors import ProjectNotFound
+from kleinkram.errors import MissionValidationError
 from kleinkram.models import File
 from kleinkram.models import Mission
 from kleinkram.models import Project
-from kleinkram.utils import is_valid_uuid4
+from kleinkram.utils import is_valid_uuid4, split_args
 
 __all__ = [
     "_get_api_version",
@@ -228,6 +231,47 @@ def _project_name_is_available(client: AuthenticatedClient, project_name: str) -
         return True
     return False
 
+def _validate_mission_created(client: AuthenticatedClient, project_id: str, mission_name: str) -> None:
+    """
+    validate that a mission is successfully created
+    """
+    mission_ids, mission_patterns = split_args([mission_name])
+    project_ids, project_patterns = split_args([project_id])
+
+    project_query = ProjectQuery(ids=project_ids, patterns=project_patterns)
+    mission_query = MissionQuery(
+        ids=mission_ids,
+        patterns=mission_patterns,
+        project_query=project_query,
+    )
+    try:
+        with tempfile.NamedTemporaryFile(suffix=".mcap", delete=False) as tmp:
+            tmp.write(b"dummy content")
+            tmp_path = Path(tmp.name)
+
+        kleinkram.core.upload(
+            client=client,
+            query=mission_query,
+            file_paths=[tmp_path],
+            verbose=False,
+        )
+
+        file_query = FileQuery(
+            ids=[],
+            patterns=[tmp_path.name],
+            mission_query=mission_query,
+        )
+        file_parsed = get_file(client, file_query)
+
+        kleinkram.core.delete_files(client=client, file_ids=[file_parsed.id])
+
+    except Exception as e:
+        raise MissionValidationError(f"Mission validation failed: {e}")
+
+    finally:
+        if tmp_path.exists():
+            tmp_path.unlink()
+
 
 def _create_mission(
     client: AuthenticatedClient,
@@ -275,6 +319,8 @@ def _create_mission(
     }
     resp = client.post(CREATE_MISSION, json=payload)
     resp.raise_for_status()
+
+    _validate_mission_created(client, str(project_id), mission_name)
 
     return UUID(resp.json()["uuid"], version=4)
 
