### 2025-09-24 2fa1ba7
fix: address ESLint errors
