### 2025-10-01 26b09b9
add remarks to getting started guide that Safari might cause issues (see #1735)
