+++ docs/development/getting-started.md
@@ -11,6 +11,7 @@ system:
 - [Git](https://git-scm.com/downloads) (for cloning the repository)
 - [Docker](https://docs.docker.com/get-docker/) (for running the project)
 - [Docker Compose](https://docs.docker.com/compose/install/) (for running the project)
+- Chrome or Firefox Browser (Safari might cause issues while developing locally)
 
 ::: info Additional Tools
 To enable code completion and linting in your IDE, you may also need a NodeJS, yarn, and python installation.
@@ -65,6 +66,11 @@ In principle the setup works similar to running the project locally, however, th
     - frontend at [http://localhost:8003](http://localhost:8003).
     - minio console at [http://localhost:9001](http://localhost:9001).
     - documentation at [http://localhost:4000](http://localhost:4000).
+
+   ::: warning
+   There are some known issues with Safari while developing locally. Please use Chrome or Firefox instead.
+   The production build does support Safari.
+   :::
 
 4. In order to enable code completion and linting in your IDE, you may need to install additional tools.
     - For JavaScript/TypeScript, you may need to install NodeJS and yarn.
