### 2025-10-01 da2d70e
fix: mission name cannot end with whitespace error message
