### 2025-10-01 7e27ec7
fix: verbose mission query error for mission delete command
