### 2025-10-01 3074d95
fix: klein mission delete abort message displays mission name if project not specified
