### 2025-10-01 63304a9
fix: kupdating metadata with non-existent tag throws verbose message
