### 2025-10-02 5435bc1
feat: add experimental support for additional file types (#1758)
