+++ cli/kleinkram/utils.py
@@ -31,6 +31,10 @@ INTERNAL_ALLOWED_CHARS = string.ascii_letters + string.digits + "_" + "-"
 SUPPORT_FILE_TYPES = [
     ".bag",
     ".mcap",
+    ".db3",
+    ".svo2",
+    ".tum",
+    ".yaml"
 ]
 
 
