+++ cli/kleinkram/cli/_upload.py
@@ -44,6 +44,10 @@ def upload(
         False,
         help="fix filenames before upload, this does not change the filenames locally",
     ),
+    experimental_datatypes: bool = typer.Option(
+       False,
+       help="allow experimental datatypes (yaml, svo2, db3, tum)"
+    ),
     ignore_missing_tags: bool = typer.Option(False, help="ignore mission tags"),
 ) -> None:
     # get filepaths
@@ -61,6 +65,18 @@ def upload(
 
     if not fix_filenames:
         for file in file_paths:
+
+            # check for experimental datatypes and throw an exception if not allowed
+            EXPERIMENTAL_DATATYPES = {".yaml", ".svo2", ".db3", ".tum"}
+
+            if not experimental_datatypes:
+                if file.suffix.lower() in EXPERIMENTAL_DATATYPES:
+                    # NOTE: Assuming a 'DatatypeNotSupported' exception exists/is defined
+                    raise FileNameNotSupported(
+                        f"Datatype '{file.suffix}' for file {file} is not supported without the "
+                        f"`--experimental-datatypes` flag. "
+                    )
+
             if not kleinkram.utils.check_filename_is_sanatized(file.stem):
                 raise FileNameNotSupported(
                     f"Only `{''.join(kleinkram.utils.INTERNAL_ALLOWED_CHARS)}` are "
