+++ cli/kleinkram/utils.py
@@ -28,14 +28,7 @@ from kleinkram.types import IdLike
 from kleinkram.types import PathLike
 
 INTERNAL_ALLOWED_CHARS = string.ascii_letters + string.digits + "_" + "-"
-SUPPORT_FILE_TYPES = [
-    ".bag",
-    ".mcap",
-    ".db3",
-    ".svo2",
-    ".tum",
-    ".yaml"
-]
+SUPPORT_FILE_TYPES = [".bag", ".mcap", ".db3", ".svo2", ".tum", ".yaml"]
 
 
 def file_paths_from_files(
