### 2025-10-03 131f8fa
fix: disambiguate klein mission delete error message
