+++ cli/kleinkram/errors.py
@@ -45,6 +45,7 @@ class FileNameNotSupported(Exception): ...
 
 class InvalidMissionMetadata(Exception): ...
 
+
 class MissionValidationError(Exception): ...
 
 
