### 2025-10-06 39f74dd
fix: project get finds exact matches instead of LIKE matches
