### 2025-10-06 ec7c32c
fix: add parameter to specify exact matching query for deleting projects by name
