### 2025-10-06 2df5b44
fix: address no inferrable types lint error
