+++ backend/src/services/action.service.ts
@@ -100,11 +100,11 @@ export class ActionService {
         data: CreateTemplateDto,
         auth: AuthHeader,
     ): Promise<ActionTemplateDto> {
-        if (!data.dockerImage.startsWith('rslethz/')) {
-            throw new ConflictException(
-                'Only images from the rslethz namespace are allowed',
-            );
-        }
+        //if (!data.dockerImage.startsWith('rslethz/')) {
+        //    throw new ConflictException(
+        //        'Only images from the rslethz namespace are allowed',
+        //    );
+        //}
         const exists = await this.actionTemplateRepository.exists({
             where: {
                 name: data.name,
@@ -140,11 +140,11 @@ export class ActionService {
         data: UpdateTemplateDto,
         auth: AuthHeader,
     ): Promise<ActionTemplateDto> {
-        if (!data.dockerImage.startsWith('rslethz/')) {
-            throw new ConflictException(
-                'Only images from the rslethz namespace are allowed',
-            );
-        }
+        //if (!data.dockerImage.startsWith('rslethz/')) {
+        //    throw new ConflictException(
+        //        'Only images from the rslethz namespace are allowed',
+        //    );
+        //}
         const template = await this.actionTemplateRepository.findOneOrFail({
             where: { uuid: data.uuid },
         });
