### 2025-10-10 95ad76d
introduce DatatypeNotSupported instea of iFileNameNotSupported
