+++ cli/kleinkram/errors.py
@@ -43,6 +43,9 @@ class FileTypeNotSupported(Exception): ...
 class FileNameNotSupported(Exception): ...
 
 
+class DatatypeNotSupported(Exception): ...
+
+
 class InvalidMissionMetadata(Exception): ...
 
 
