+++ cli/setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = kleinkram
-version = 0.51.0
+version = 0.52.0
 description = give me your bags
 long_description = file: README.md
 long_description_content_type = text/markdown
