### 2025-10-10 973e5e2
feat: missions get command returns files count and files size
