+++ backend/src/services/mission.service.ts
@@ -32,7 +32,7 @@ import {
 } from '../serialization';
 import { TagService } from './tag.service';
 import { UserService } from './user.service';
-import { addMissionFilters, addProjectFilters, addSort } from './utilities';
+import { addMissionFilters, addProjectFilters, addSort, addFileStats } from './utilities';
 
 import { SortOrder } from '@common/api/types/pagination';
 
@@ -150,7 +150,7 @@ export class MissionService {
         return missionEntityToDtoWithFiles(mission);
     }
 
-    async findMany(
+        async findMany(
         projectUuids: string[],
         projectPatterns: string[],
         missionUuids: string[],
@@ -162,24 +162,25 @@ export class MissionService {
         take: number,
         userUuid: string,
     ): Promise<MissionsDto> {
-        let query = this.missionRepository
+        let idQuery = this.missionRepository
             .createQueryBuilder('mission')
-            .leftJoinAndSelect('mission.project', 'project')
-            .leftJoinAndSelect('mission.creator', 'creator')
-            .leftJoinAndSelect('mission.tags', 'tag')
-            .leftJoinAndSelect('tag.tagType', 'tagType');
+            .select('mission.uuid')
+            .leftJoin('mission.project', 'project')
+            .leftJoin('mission.creator', 'creator')
+            .leftJoin('mission.tags', 'tag')
+            .leftJoin('tag.tagType', 'tagType');
 
-        query = addAccessConstraintsToMissionQuery(query, userUuid);
+        idQuery = addAccessConstraintsToMissionQuery(idQuery, userUuid);
 
-        query = addProjectFilters(
-            query,
+        idQuery = addProjectFilters(
+            idQuery,
             this.projectRepository,
             projectUuids,
             projectPatterns,
         );
 
-        query = addMissionFilters(
-            query,
+        idQuery = addMissionFilters(
+            idQuery,
             this.missionRepository,
             missionUuids,
             missionPatterns,
@@ -187,11 +188,70 @@ export class MissionService {
         );
 
         if (sortBy !== undefined) {
-            query = addSort(query, FIND_MANY_SORT_KEYS, sortBy, sortOrder);
+            idQuery = addSort(idQuery, FIND_MANY_SORT_KEYS, sortBy, sortOrder);
         }
 
-        query.take(take).skip(skip);
-        const [missions, count] = await query.getManyAndCount();
+        // Get distinct mission UUIDs
+        idQuery.groupBy('mission.uuid');
+
+        // Get count before pagination
+        const count = await idQuery.getCount();
+        idQuery.take(take).skip(skip);
+
+        const missionIds = await idQuery.getRawMany();
+
+        if (missionIds.length === 0) {
+            return {
+                data: [],
+                count,
+                skip,
+                take,
+            };
+        }
+
+        let dataQuery = this.missionRepository
+            .createQueryBuilder('mission')
+            .leftJoinAndSelect('mission.project', 'project')
+            .leftJoinAndSelect('mission.creator', 'creator')
+            .leftJoinAndSelect('mission.tags', 'tag')
+            .leftJoinAndSelect('tag.tagType', 'tagType')
+            .where('mission.uuid IN (:...missionIds)', {
+                missionIds: missionIds.map(m => m.mission_uuid),
+            });
+
+        if (sortBy !== undefined) {
+            dataQuery = addSort(dataQuery, FIND_MANY_SORT_KEYS, sortBy, sortOrder);
+        }
+
+        dataQuery = addFileStats(dataQuery);
+
+        const result = await dataQuery.getRawAndEntities();
+        const missions = result.entities;
+        const rawResults = result.raw;
+
+        // Create a map for quick lookup of file stats by mission UUID
[…]
