+++ backend/src/services/utilities.ts
@@ -203,20 +203,20 @@ export const addMissionCount = (
 };
 
 export const addFileStats = (
-  query: SelectQueryBuilder<Mission>,
+    query: SelectQueryBuilder<Mission>,
 ): SelectQueryBuilder<Mission> => {
     query
         .addSelect((subQuery) => {
             return subQuery
-            .select('COUNT(file.uuid)', 'count')
-            .from('file_entity', 'file')
-            .where('file."missionUuid" = mission.uuid');
+                .select('COUNT(file.uuid)', 'count')
+                .from('file_entity', 'file')
+                .where('file."missionUuid" = mission.uuid');
         }, 'fileCount')
         .addSelect((subQuery) => {
             return subQuery
-            .select('COALESCE(SUM(file.size), 0)', 'sum')
-            .from('file_entity', 'file')
-            .where('file."missionUuid" = mission.uuid');
+                .select('COALESCE(SUM(file.size), 0)', 'sum')
+                .from('file_entity', 'file')
+                .where('file."missionUuid" = mission.uuid');
         }, 'fileSize');
 
     return query;
