### 2025-10-11 cf9c241
feat: Add transparent zstd compression support for rosbag uploads
This adds support for uploading zstd-compressed rosbag files (.zst/.zstd)
without requiring manual decompression. Only zstd is supported due to its
superior performance (5-10x faster than gzip/bzip2).

## Why Only Zstd?

Performance comparison for typical 1.8GB uncompressed rosbag:
- **zstd**: ~2 seconds decompression (1000 MB/s) ✅
- **gzip**: ~6-18 seconds (100-300 MB/s) ❌ Too slow
- **bzip2**: ~36 seconds (20-50 MB/s) ❌ Way too slow

Result: Simpler implementation, optimal user experience.

## Changes

### New module: `cli/kle
[…]
