+++ cli/tests/test_compression.py
@@ -1,0 +1,171 @@
+"""
+Tests for zstd compression support in kleinkram CLI.
+
+Why only zstd?
+- 5-10x faster than gzip/bzip2
+- Excellent compression ratios
+- Industry standard for ROS bags
+"""
+
+from __future__ import annotations
+
+import tempfile
+from pathlib import Path
+
+import pytest
+
+from kleinkram import compression, utils
+
+
+class TestCompressionDetection:
+    """Test zstd compression detection."""
+
+    def test_detect_zst(self):
+        assert compression.detect_compression(Path("file.mcap.zst")) == compression.CompressionType.ZSTD
+
+    def test_detect_zstd(self):
+        assert compression.detect_compression(Path("file.bag.zstd")) == compression.CompressionType.ZSTD
+
+    def test_detect_no_compression(self):
+        assert compression.detect_compression(Path("file.mcap")) is None
+        assert compression.detect_compression(Path("file.bag")) is None
+
+    def test_case_insensitive(self):
+        assert compression.detect_compression(Path("file.MCAP.ZST")) == compression.CompressionType.ZSTD
+        assert compression.detect_compression(Path("file.Bag.ZSTD")) == compression.CompressionType.ZSTD
+
+
+class TestUncompressedName:
+    """Test getting uncompressed filenames."""
+
+    def test_remove_zst(self):
+        assert compression.get_uncompressed_name(Path("data.mcap.zst")) == "data.mcap"
+
+    def test_remove_zstd(self):
+        assert compression.get_uncompressed_name(Path("data.bag.zstd")) == "data.bag"
+
+    def test_no_compression(self):
+        assert compression.get_uncompressed_name(Path("data.mcap")) == "data.mcap"
+
+    def test_complex_name(self):
+        assert compression.get_uncompressed_name(Path("my_data_2024.mcap.zst")) == "my_data_2024.mcap"
+
+
+class TestFileValidation:
+    """Test that zstd-compressed files pass validation."""
+
+    def test_check_file_path_compressed(self):
+        """Test that zstd-compressed files with valid base types are accepted."""
+        # Create temporary compressed file
+        with tempfile.NamedTemporaryFile(suffix=".mcap.zst", delete=False) as f:
+            temp_path = Path(f.name)
+
+        try:
+            # Should not raise an exception
+            utils.check_file_path(temp_path)
+        finally:
+            temp_path.unlink()
+
+    def test_get_filename_strips_compression(self):
+        """Test that get_filename removes zstd compression extension."""
+        assert utils.get_filename(Path("data.mcap.zst")) == "data.mcap"
+        assert utils.get_filename(Path("data.bag.zstd")) == "data.bag"
+        assert utils.get_filename(Path("data.mcap")) == "data.mcap"
+
+    def test_sanitize_compressed_filename(self):
+        """Test filename sanitization works with zstd-compressed files."""
+        # Spaces should be replaced with underscores
+        assert utils.get_filename(Path("my data.mcap.zst")) == "my_data.mcap"
+        assert utils.get_filename(Path("test_file.bag.zstd")) == "test_file.bag"
+
+
+class TestCompressionExtensionsInUtils:
+    """Test that COMPRESSION_EXTENSIONS is available in utils module."""
+
+    def test_compression_extensions_defined(self):
+        assert hasattr(utils, 'COMPRESSION_EXTENSIONS')
+        assert isinstance(utils.COMPRESSION_EXTENSIONS, list)
+        assert '.zst' in utils.COMPRESSION_EXTENSIONS
+        assert '.zstd' in utils.COMPRESSION_EXTENSIONS
+
+
+# Integration-style tests that would require actual compression/decompression
[…]
