### 2025-10-11 b57d543
chore(cli): bump version to 0.51.1
