### 2025-10-13 1bd80a8
feat: define environment variable for docker hub namespace restrictions
