+++ common/environment.ts
@@ -190,4 +190,12 @@ export default {
     get VITE_USE_FAKE_OAUTH_FOR_DEVELOPMENT(): boolean {
         return asBoolean(process.env['VITE_USE_FAKE_OAUTH_FOR_DEVELOPMENT']);
     },
+
+    /**
+     * @returns Docker Hub namespace for image validation (optional)
+     * @example rslethz/
+     */
+    get DOCKER_HUB_NAMESPACE(): string {
+        return process.env['VITE_DOCKER_HUB_NAMESPACE'] ?? '';
+    },
 };
