### 2025-10-13 6f505c9
fix: get_project has exact_match param set to True where needed
