+++ .github/workflows/eslint.yml
@@ -14,7 +14,7 @@ jobs:
             - uses: actions/checkout@v5
 
             - name: Use Node.js
-              uses: actions/setup-node@v5
+              uses: actions/setup-node@v6
               with:
                   node-version: '22'
                   package-manager-cache: false
