### 2025-11-11 a9fafd9
add backend support for additional file types #1835
