### 2025-11-11 75cf150
add missing function return types
