+++ backend/src/services/file.service.ts
@@ -24,6 +24,7 @@ import {
 import { InjectRepository } from '@nestjs/typeorm';
 import jwt from 'jsonwebtoken';
 import {
+    Brackets,
     DataSource,
     FindOptionsSelect,
     FindOptionsWhere,
@@ -31,6 +32,7 @@ import {
     In,
     MoreThan,
     Repository,
+    SelectQueryBuilder,
 } from 'typeorm';
 import { fileEntityToDto, fileEntityToDtoWithTopic } from '../serialization';
 import {
@@ -51,7 +53,6 @@ import { StorageOverviewDto } from '@common/api/types/storage-overview.dto';
 import { redis } from '@common/consts';
 import Category from '@common/entities/category/category.entity';
 import QueueEntity from '@common/entities/queue/queue.entity';
-import Tag from '@common/entities/tag/tag.entity';
 import TagType from '@common/entities/tagType/tag-type.entity';
 import User from '@common/entities/user/user.entity';
 import {
@@ -179,6 +180,17 @@ export class FileService implements OnModuleInit {
         };
     }
 
+    /**
+     * Finds and paginates files based on a comprehensive set of filters.
+     *
+     * This method uses a two-query approach to correctly handle pagination with
+     * complex joins and groupings (required for 'matchAll' topics and tags):
+     * 1. The first query applies all filters and retrieves *only* the UUIDs
+     * of the matching files for the requested page, along with the *total count*
+     * of all matching files (pre-pagination).
+     * 2. The second query fetches the full file entities (with relations) for
+     * the UUIDs retrieved in the first query.
+     */
     async findFiltered(
         fileName: string,
         projectUUID: string,
@@ -186,7 +198,7 @@ export class FileService implements OnModuleInit {
         startDate: Date | undefined,
         endDate: Date | undefined,
         topics: string,
-        andOr: boolean,
+        matchAllTopics: boolean,
         fileTypes: string,
         tags: Record<string, any>,
         userUUID: string,
@@ -198,208 +210,66 @@ export class FileService implements OnModuleInit {
         const user = await this.userRepository.findOneOrFail({
             where: { uuid: userUUID },
         });
-        // Start building your query with basic filters
-        let query = this.fileRepository
-            .createQueryBuilder('file')
-            .select('file.uuid')
-            .leftJoin('file.mission', 'mission')
-            .leftJoin('file.topics', 'topic')
-            .leftJoin('mission.project', 'project')
-            .offset(skip)
-            .limit(take)
-            .orderBy(sort, sortOrder);
 
-        // ADMIN user have access to all files, all other users have access to files based on their access
+        // Start building the query to fetch *only* IDs
+        let idQuery = this.fileRepository
+            .createQueryBuilder('file')
+            .select('file.uuid') // Select only the UUID
+            .leftJoin('file.mission', 'mission')
+            .leftJoin('mission.project', 'project')
+            .leftJoin('file.topics', 'topic'); // Joined for filtering
+
+        // ADMIN users see all, others are constrained
         if (user.role !== UserRole.ADMIN) {
-            query = addAccessConstraints(query, userUUID);
+            idQuery = addAccessConstraints(idQuery, userUUID);
         }
 
-        // Apply filters for fileName, projectUUID, and date
+        // Apply simple filters
         if (fileName) {
             logger.debug(`Filtering files by filename: ${fileName}`);
-            query.andWhere('file.filename LIKE :fileName', {
+            idQuery.andWhere('file.filename LIKE :fileName', {
                 fileName: `%${fileName}%`,
             });
         }
 
-        if (fileTypes) {
-            const requestedTypes = fileTypes.split(',');
-
-            // Check for 'ALL' case-insensitively
[…]
