### 2025-11-11 9879895
remove outdated filesOfMission endpoint and use filtered endpoint instead
