+++ frontend/src/services/query-handler.ts
@@ -5,7 +5,7 @@ import { RouteLocationNormalizedLoaded, Router } from 'vue-router';
 const DEFAULT_SORT = { sortBy: 'name', descending: false };
 const DEFAULT_PAGINATION = { page: 1, rowsPerPage: 20 };
 const DEFAULT_SEARCH = {};
-const DEFAULT_FILE_TYPE = FileType.ALL;
+const DEFAULT_FILE_TYPES: FileType[] = Object.values(FileType);
 
 /**
  * Class passed around to handle the query state
@@ -19,9 +19,10 @@ export class QueryHandler {
     missionUuid?: string;
     projectUuid?: string;
     searchParams: Record<string, string>;
-    fileType?: FileType;
     rowsNumber: number;
     categories: string[];
+
+    fileTypes: FileType[];
 
     constructor(
         page: number = DEFAULT_PAGINATION.page,
@@ -31,7 +32,7 @@ export class QueryHandler {
         projectUuid?: string,
         missionUuid?: string,
         searchParameters?: Record<string, string>,
-        fileType?: FileType,
+        fileTypes?: FileType[],
         categories?: string[],
     ) {
         this.page = page;
@@ -41,9 +42,10 @@ export class QueryHandler {
         this.projectUuid = projectUuid ?? '';
         this.missionUuid = missionUuid ?? '';
         this.searchParams = searchParameters ?? DEFAULT_SEARCH;
-        this.fileType = fileType === undefined ? FileType.ALL : fileType;
         this.categories = categories ?? [];
         this.rowsNumber = 0;
+
+        this.fileTypes = fileTypes ?? [...DEFAULT_FILE_TYPES];
     }
 
     get skip(): number {
@@ -105,8 +107,8 @@ export class QueryHandler {
         }
     }
 
-    setFileType(fileType: FileType): void {
-        this.fileType = fileType;
+    setFileTypes(fileTypes: FileType[]): void {
+        this.fileTypes = fileTypes;
         this.resetPagination();
     }
 
@@ -126,11 +128,11 @@ export class QueryHandler {
         this.page = DEFAULT_PAGINATION.page;
     }
 
-    get isListingMissions() {
+    get isListingMissions(): boolean {
         return !!this.projectUuid && !this.missionUuid;
     }
 
-    get isListingFiles() {
+    get isListingFiles(): boolean {
         return !!this.missionUuid;
     }
 
@@ -139,13 +141,23 @@ export class QueryHandler {
      * Part of each query key
      * ---------------------------------------------
      */
-    get queryKey() {
+    get queryKey(): {
+        searchParams: Record<string, string>;
+        skip: number;
+        take: number;
+        sortBy: string;
+        descending: boolean;
+        fileTypes: FileType[];
+        categories: string[];
+    } {
         return {
             searchParams: this.searchParams,
             skip: this.skip,
             take: this.take,
             sortBy: this.sortBy,
             descending: this.descending,
+            fileTypes: this.fileTypes,
+            categories: this.categories,
         };
     }
 }
@@ -166,7 +178,7 @@ export class QueryURLHandler extends QueryHandler {
         projectUuid?: string,
         missionUuid?: string,
         searchParameters?: typeof DEFAULT_SEARCH,
-        fileType?: FileType,
+        fileTypes?: FileType[],
         categories?: string[],
     ) {
         super(
@@ -177,7 +189,7 @@ export class QueryURLHandler extends QueryHandler {
             projectUuid,
             missionUuid,
             searchParameters,
-            fileType,
+            fileTypes,
             categories,
         );
         if (router) {
@@ -256,8 +268,9 @@ export class QueryURLHandler extends QueryHandler {
         });
     }
 
[…]

[…]
