### 2025-11-11 7d34986
introduce HealthStatus enum
