### 2025-11-11 0a587a5
allow uploading files via web gui (#1758)
