### 2025-11-11 2eef227
use enum instead of hardcoded strings
