### 2025-11-11 fce9f10
fix #1884
