### 2025-11-11 aeed215
first step for #1759: link to docs on action page
