### 2025-11-12 613a6f5
add missing return types
