### 2025-11-12 506a3b4
update docs for minio
