### 2025-11-12 23adb8c
implement #1868
