### 2025-11-15 8c438a1
implement #1834
