### 2025-11-15 e11d385
general availability for .db3", ".svo2", ".tum", ".yaml", ".yml"
