### 2025-11-15 325bed3
avoid 500 if no project is found
