+++ cli/kleinkram/utils.py
@@ -33,7 +33,7 @@ EXPERIMENTAL_FILE_TYPES = []
 
 
 def file_paths_from_files(
-        files: Sequence[File], *, dest: Path, allow_nested: bool = False
+    files: Sequence[File], *, dest: Path, allow_nested: bool = False
 ) -> Dict[Path, File]:
     """\
     determines the destinations for a sequence of `File` objects,
@@ -230,7 +230,7 @@ def format_bytes(size_bytes: int | float, speed: bool = False) -> str:
     power = math.floor(math.log(size_bytes, 1000))
     unit_index = min(power, len(units) - 1)
 
-    value = size_bytes / (1000 ** unit_index)
+    value = size_bytes / (1000**unit_index)
 
     unit_suffix = "/s" if speed else ""
     return f"{value:.2f} {units[unit_index]}{unit_suffix}"
