### 2025-11-15 e190633
fix exact match search by name
