### 2025-11-15 b552a7e
add missing types
