+++ backend/src/services/utilities.ts
@@ -4,7 +4,12 @@ import Mission from '@common/entities/mission/mission.entity';
 import Project from '@common/entities/project/project.entity';
 import { MethodNotAllowedException } from '@nestjs/common';
 import { isValid, parseISO } from 'date-fns';
-import { Brackets, Repository, SelectQueryBuilder } from 'typeorm';
+import {
+    Brackets,
+    ObjectLiteral,
+    Repository,
+    SelectQueryBuilder,
+} from 'typeorm';
 import { v4 as uuidv4 } from 'uuid';
 
 export const stringToBoolean = (value: string): boolean | undefined => {
@@ -42,7 +47,7 @@ export const stringToLocation = (value: string): string | undefined => {
     return value;
 };
 
-export const convertGlobToLikePattern = (glob: string) => {
+export const convertGlobToLikePattern = (glob: string): string => {
     return glob
         .replaceAll('%', String.raw`\%`)
         .replaceAll('_', String.raw`\_`)
@@ -194,9 +199,9 @@ export const getFilteredFileIdSubQuery = (
     return query;
 };
 
-export const addMissionCount = (
-    query: SelectQueryBuilder<any>,
-): SelectQueryBuilder<any> => {
+export const addMissionCount = <T extends ObjectLiteral>(
+    query: SelectQueryBuilder<T>,
+): SelectQueryBuilder<T> => {
     query.loadRelationCountAndMap(
         'project.missionCount',
         'project.missions',
@@ -226,10 +231,10 @@ export const addFileStats = (
     return query;
 };
 
-export const addProjectCreatorFilter = (
-    query: SelectQueryBuilder<any>,
+export const addProjectCreatorFilter = <T extends ObjectLiteral>(
+    query: SelectQueryBuilder<T>,
     creatorUuid: string | undefined,
-): SelectQueryBuilder<any> => {
+): SelectQueryBuilder<T> => {
     if (creatorUuid) {
         query.andWhere('project.creator.uuid = :creatorUuid', {
             creatorUuid,
@@ -239,13 +244,13 @@ export const addProjectCreatorFilter = (
     return query;
 };
 
-export const addProjectFilters = (
-    query: SelectQueryBuilder<any>,
+export const addProjectFilters = <T extends ObjectLiteral>(
+    query: SelectQueryBuilder<T>,
     projectRepository: Repository<Project>,
     projectIds: string[],
     projectPatterns: string[],
     exactMatch = false,
-): SelectQueryBuilder<any> => {
+): SelectQueryBuilder<T> => {
     if (projectIds.length > 0 || projectPatterns.length > 0) {
         const projectLikePatterns = projectPatterns.map((element) =>
             convertGlobToLikePattern(element),
@@ -266,13 +271,13 @@ export const addProjectFilters = (
     return query;
 };
 
-export const addMissionFilters = (
-    query: SelectQueryBuilder<any>,
+export const addMissionFilters = <T extends ObjectLiteral>(
+    query: SelectQueryBuilder<T>,
     missionRepository: Repository<Mission>,
     missionIds: string[],
     missionPatterns: string[],
     missionMetadata: Record<string, string>,
-): SelectQueryBuilder<any> => {
+): SelectQueryBuilder<T> => {
     if (
         missionIds.length > 0 ||
         missionPatterns.length > 0 ||
@@ -297,12 +302,12 @@ export const addMissionFilters = (
     return query;
 };
 
-export const addFileFilters = (
-    query: SelectQueryBuilder<any>,
+export const addFileFilters = <T extends ObjectLiteral>(
+    query: SelectQueryBuilder<T>,
     fileRepository: Repository<File>,
[…]
