### 2025-11-15 f8d6d88
avoid 500 on EntityNotFoundError (#1390)
