### 2025-11-15 9f14a58
return meaningfully error (#1390)
