### 2025-11-17 leggedrobotics/kleinkram#1894 @c-joly
Actions: Cannot run a custom action
JOSS context: openjournals/joss-reviews#8636
cc: @boisgera

I tried to run a custom action as suggested by the doc (the one that is supposed to list the files of the mission):
  - I successfully created a container that starts with the defined entry point under my docker hub namespace
  - I successfully submitted it from the Kleinkram interface (see attached screenshot)
  - However, it fails to download the mcap file (access denied, see attached screenshot)

I slightly changed the `entrypoint.sh` file to display the value of the key `KLEINKRAM_API_KEY` which is used for logging. It displays `***` (third line in the screenshot of the log)… Is it possible that the error is due to a bad  value in `KLEINKRAM_API_KEY`? And in that case, how to solve it?

FYI, I am using the  **Kleinkram Admin User Nr. 1**  account (without a *Dev-login* account, I cannot create any project).
