### 2025-11-17 leggedrobotics/kleinkram#1895 @c-joly
Impossible to create a project with a non-admin account
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636
cc: @boisgera

I tried to login with my Google account but I don't have the right to create any project. @boisgera suggested to me to change `backend/access_config.json` to make Kleinkram accepts gmail domain, but it did not work. Here is my new version of the file:

```json
{
    "emails": [
        {
            "email": "gmail.com",
            "access_groups": ["00000000-0000-0000-0000-000000000000"]
        }
    ],
    "access_groups": [
        {
            "name": "cyril",
            "uuid": "00000000-0000-0000-0000-000000000000",
            "rights": 30
        }
    ]
}
```

Can you tell me what I am doing wrong?
