### 2025-11-17 leggedrobotics/kleinkram#1894 @wp99cp
[978fd19](https://github.com/leggedrobotics/kleinkram/commit/978fd19fc92bf05965089528f02e6324bf98c702) fixes the issue. The fix is available on the `dev` branch and will be included in the 0.53.0 release of kleinkram. Due to a small difference in the SQL schema this had not effected our production istance at [(dataset.leggedrobotics.com) ](https://datasets.leggedrobotics.com/) but only the dev system and local dev enviroment.
