### 2025-11-17 leggedrobotics/kleinkram#1895 @wp99cp
At the moment we do not support changes to `backend/access_config.json` with an existing database state. Make sure to start with a fresh database state after applying changes to `access_config.json`, i.e. do the following:

1) update the `backend/access_config.json` file as described.
2) disable DB Seeding in `.env` by setting `SEED=false`
3) run `docker compose down --volumes`
4) start the application `docker compose up --build --watch`

Let me know if that helped.

I can now log in with my `***@gmail.com` address and get added sucessfully to an affiliation group called `cyril`. This can be checked under http://localhost:8003/access-groups by selecting `Affiliation Group` in the `Groups`dropdown.

Now I can also create projects when logged in with my Gmail Account:
