### 2025-11-18 leggedrobotics/kleinkram#1895 @wp99cp
You can keep `SEED` set to `false`. That shouldn't effect anything in your case.

Enabling the `SEED` variable adds dummy data on the application's first startup and should only be used with local dev accounts (`fake-oauth`). With seeding enabled, it is super easy to set up a reproducible test environment while developing the application.
