### 2025-11-19 leggedrobotics/kleinkram#1894 @wp99cp
!! Never mind, I can reproduce the issue on a clean install !!

By using the following steps

1) clone kleinkram into a new folder
   `cd /tmp`
   `git clone https://github.com/leggedrobotics/kleinkram.git`
   `cd kleinkram`
2) Clear any old data
    `docker compose down --volumes`
3) Disable Seeding
    by setting `SEED=false` inside the `.env`
4) Start the Application
    `docker compose up --build`
5) In a separate terminal, install the CLI from source
    `cd /tmp/kleinkram/cli`
    `virtualenv .venv`
    `source .venv/bin/activate`
    `pip install -e . -r requirements-dev.txt`
6) Authenticate the CLI
    `klein endpoint local`
    `klein login`  --> select the Admin Account in the Fake-oAuth screen
7) Create a Project and Mission
    `klein project create -p testProject -d "Just a Test Project for running actions"`
    `echo "123" > test.yml`
    `klein upload -p testProject -m testMission --create test.yml`
8) Build a test Action
    `cd /tmp/kleinkram/examples/actions`
    `docker build -t <your-docker-hub-account>/test_action .`
     `docker push <your-docker-hub-account>/test_action`
9) Define the Action Template using the Web GUI and save it as a template

   make sure to use your image `<your-docker-hub-account>/test_action`
   (we have an open issue to support locally build images, #842)

10) Start the action using the CLI
    `klein action list`
    `klein action run test_action -p testProject -m testMission --follow`
