### 2025-11-24 cc96bc9
chore: cleanup version checking middleware
