+++ backend/src/routing/middlewares/version-checker-middleware.service.ts
@@ -1,12 +1,27 @@
 import { Injectable, NestMiddleware } from '@nestjs/common';
+import { HttpException } from '@nestjs/common/exceptions/http.exception';
 import { NextFunction, Request, Response } from 'express';
 import { appVersion } from '../../app-version';
 import logger from '../../logger';
 
 /**
+ * Configuration for forbidden version ranges.
+ * Supported patterns:
+ * - `<x.y.z`: Block all versions strictly less than x.y.z.
+ * - `x.y.*`: Block all versions in the x.y family (wildcard *).
+ * - `x.y.z`: Block an exact version match.
  *
- * A nest middleware that resolves the user from the API key in the request.
+ */
+const FORBIDDEN_RANGES = ['<0.55.1'];
+
+/**
+ * Enforces client version compatibility by inspecting the request headers.
  *
+ * This middleware extracts the `kleinkram-client-version`, sanitizes it,
+ * and validates it against a defined blocklist (`FORBIDDEN_RANGES`).
+ *
+ * If the version is missing, invalid, or forbidden, the request is rejected
+ * with a 426 (Upgrade Required) status.
  */
 @Injectable()
 export class VersionCheckerMiddlewareService implements NestMiddleware {
@@ -16,12 +31,6 @@ export class VersionCheckerMiddlewareService implements NestMiddleware {
             | undefined;
 
         const requestPath = request.originalUrl;
-
-        // ignore auth endpoints
-        if (requestPath.startsWith('/auth/')) {
-            next();
-            return;
-        }
 
         // strip away everything after .dev
         if (clientVersion !== undefined) {
@@ -44,10 +53,7 @@ export class VersionCheckerMiddlewareService implements NestMiddleware {
             return;
         }
 
-        // forbidden client versions: allows for the following notations
-        const forbiddenClientVersions = ['<0.55.1'];
-
-        if (this.isVersionForbidden(clientVersion, forbiddenClientVersions)) {
+        if (this.isVersionForbidden(clientVersion, FORBIDDEN_RANGES)) {
             this.rejectRequest(response, clientVersion);
             return;
         }
@@ -55,19 +61,35 @@ export class VersionCheckerMiddlewareService implements NestMiddleware {
         next();
     }
 
+    /**
+     * Terminates the request with a 426 Upgrade Required error.
+     *
+     * Side Effect:
+     * - Sets the `kleinkram-version` header to the current `appVersion`
+     * so the client knows which version the server is running.
+     *
+     * @throws HttpException (426)
+     */
     private rejectRequest(response: Response, clientVersion: string): void {
-        // we need to manually set the version header here
+        // Set the version header to inform the client about the required version
         response.header('kleinkram-version', appVersion);
 
-        // reject request with 426
-        response.status(426).json({
-            statusCode: 426,
-            message: `Client version ${clientVersion} is not a valid version.`,
-        });
-
-        response.send();
+        throw new HttpException(
+            {
+                statusCode: 426,
+                message: `Client version ${clientVersion} is not supported. Please upgrade to at least version ${appVersion}.`,
+                error: 'Upgrade Required',
+            },
+            426, // HTTP 426 Upgrade Required
+        );
     }
 
+    /**
+     * Checks if a specific client version matches any rule in the forbidden list.
+     *
+     * @param clientVersion - The sanitized client version (e.g. "0.55.0")
+     * @param forbiddenVersions - Array of rules (e.g. ["<0.55.1", "1.0.x"])
+     */
     private isVersionForbidden(
         clientVersion: string,
         forbiddenVersions: string[],
@@ -78,7 +100,7 @@ export class VersionCheckerMiddlewareService implements NestMiddleware {
[…]
