### 2025-11-24 e3da785
feat: add support for foxglove preview links (#1936)
