+++ backend/src/app.module.ts
@@ -14,6 +14,7 @@ import { ActionModule } from './endpoints/action/action.module';
 import { AuthModule } from './endpoints/auth/auth.module';
 import { CategoryModule } from './endpoints/category/category.module';
 import { FileModule } from './endpoints/file/file.module';
+import { FoxgloveModule } from './endpoints/integrations/foxglove.module';
 import { MissionModule } from './endpoints/mission/mission.module';
 import { ProjectModule } from './endpoints/project/project.module';
 import { QueueModule } from './endpoints/queue/queue.module';
@@ -64,6 +65,7 @@ import { AccessGroupConfig } from './types/access-group-config';
             inject: [ConfigService],
         }),
         FileModule,
+        FoxgloveModule,
         ProjectModule,
         TopicModule,
         MissionModule,
@@ -99,6 +101,7 @@ export class AppModule implements NestModule {
             .apply(VersionCheckerMiddlewareService)
             .exclude(
                 '/auth/(.*)', // excludes auth endpoints
+                '/integrations/(.*)', // excludes integration endpoints
             )
             .forRoutes('*');
     }
