+++ common/frontend_shared/enum.ts
@@ -167,6 +167,7 @@ export enum FileEventType {
     TOPICS_EXTRACTED = 'TOPICS_EXTRACTED',
     FILE_CONVERTED = 'FILE_CONVERTED',
     FILE_CONVERTED_FROM = 'FILE_CONVERTED_FROM',
+    FOXGLOVE_URL_GENERATED = 'FOXGLOVE_URL_GENERATED',
 
     // User Actions
     DOWNLOADED = 'DOWNLOADED',
