### 2025-11-24 761abbd
fix: fox CORS issues related to foxglove links (#1936)
