### 2025-11-24 f6d64d2
fix: fox CORS issues related to foxglove links (#1936)
