+++ backend/src/main.ts
@@ -48,11 +48,7 @@ async function bootstrap(): Promise<void> {
         }),
     );
     app.enableCors({
-        origin: [
-            environment.FRONTEND_URL,
-            environment.DOCS_URL,
-            'https://app.foxglove.dev/',
-        ],
+        origin: [environment.FRONTEND_URL, environment.DOCS_URL],
         methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
         preflightContinue: false,
         credentials: true,
