### 2025-11-24 4ca511c
fix: update CORS policy to use null as default value
