### 2025-11-24 #8636 @boisgera
> Hi everyone,
>
> Could we proceed on this here?
>
> Best regards, Jonas

Hi @JonasFrey96,

I just had a discussion with @c-joly about the state of the review. At this stage we can't in good faith check some mandatory requirements (in the Functionality and Documentation sections) of the JOSS review process. At this stage @c-joly and I have already spent an unusual amount of time trying to reproduce locally the documented Kleinkram features, but many issues required some changes in the software or the communication of informations that were not publicly available. (Every time @wp99cp was really reactive to theses issues and I'd like thank him for that! 🙏).

At this stage, I have (for unknown reasons) a regression that prevents me to create new missions/projects in my local instance of Kleinkram, a thing that I managed to do previously ; therefore cannot take the necessary steps to test the software features. OTOH @c-joly is making progress but did not manage to execute his own dummy action yet to his satisfaction.

Beyond the checklist and more broadly: I am sure that this project is useful to you internally and there was clearly a lot of work behind it. But given the limited success that @c-joly and I to use the project despite our efforts and the guidance of @wp99cp, I doubt that in its current state the project can be broadly used by external teams without access to the deep knowledge that you have about the stack. It may be a sign that we should suspend the submission until the project is ready to reach a wider public (it may *also* be a sign that @c-joly and I are not the appropriate rewievers, of course!)

Since I am a reviewer **and** an editor on this review, I'd like to get some independent assessment of the situation. Therefore, I'll ask my fellow JOSS editors what they think of this situation and I'll report here shortly,

Best regards,

Sébastien (@boisgera)
