### 2025-11-24 d386ab3
fix: correct typo in WAIT_BEFORE_DELETION constant
