+++ .github/workflows/production_build.yml
@@ -10,7 +10,7 @@ jobs:
         runs-on: ubuntu-latest
 
         steps:
-            - uses: actions/checkout@v5
+            - uses: actions/checkout@v6
             - name: Login to DockerHub
               uses: docker/login-action@v3
               with:
