+++ .github/workflows/test.yml
@@ -18,7 +18,7 @@ jobs:
             pull-requests: write # Allows adding comments to PRs
 
         steps:
-            - uses: actions/checkout@v5
+            - uses: actions/checkout@v6
 
             - name: Use Node.js
               uses: actions/setup-node@v6
