+++ .github/workflows/eslint.yml
@@ -11,7 +11,7 @@ jobs:
         runs-on: [self-hosted, linux]
 
         steps:
-            - uses: actions/checkout@v5
+            - uses: actions/checkout@v6
 
             - name: Use Node.js
               uses: actions/setup-node@v6
