+++ .github/workflows/check_version_number.yml
@@ -13,7 +13,7 @@ jobs:
 
         steps:
             - name: Checkout repository
-              uses: actions/checkout@v5
+              uses: actions/checkout@v6
 
             - uses: actions/setup-node@v6
               with:
