### 2025-11-25 e76550d
fix: add fallback to headless until #1640 is fixed
