### 2025-11-25 f87b86c
feat: implement #1935
