### 2025-11-25 d389c12
fix: add /integrations/ to allowed endpoints for unauthenticated users
