### 2025-11-25 479a036
fix: typo in list files by project name
