### 2025-11-25 ecc8d2f
chore: migrate /file to /files
