### 2025-11-25 leggedrobotics/kleinkram#1959 @c-joly
CLI mismatch in example action
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636
cc: @boisgera

Hi, I finally succeeded to run a custom action from my Google account.

However, with the default procedure, the example action (I used the one provided in the repository), I have an error of mismatch version between CLI and server.
The CLI version on the action container is `0.54.0.dev20251118141514` whereas the CLI version on my computer is 0.55.2 (same as the server).

I figured out that the command `RUN pip install kleinkram --pre --force-reinstall && pip install "httpx==0.26.0"` of the `Dockerfile` was the issue. I removed the `--pre` option (which looks to force the download of a version of the dev branch), and then the docker image was built with the good version of CLI. Can you update the example Dockerfile in the Github project.

By the way, I was wondering how to restart an action that is finished from the GUI? I saw an option "manually triggered" when I created the action, but I did not find any button to trigger it again (or maybe I missed it?)
