### 2025-11-25 leggedrobotics/kleinkram#1960 @c-joly
Problem in downloading file in action: failed hash check
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636%E2%80%A8
cc: @boisgera

I figured out that the example action a problem when downloading a file: the log shows `downloaded /out/test.yaml but failed hash check` (see screenshot).
Is it normal?
