### 2025-11-25 91793f8
fix: build sample action with latest prod cli version (#1959)
