### 2025-11-25 leggedrobotics/kleinkram#1959 @wp99cp
In addition, last week I added #842, which allows you to use locally built and tagged images in a local dev instance without pushing them to a Docker registry. A full guide can be found here: https://github.com/leggedrobotics/kleinkram/issues/1894#issuecomment-3551609484
