### 2025-11-26 a9b3692
fix: missed endpoint change
