### 2025-11-26 417271a
feat: add endpoint to check if action template name is valid
