### 2025-11-26 e6182c1
feat: add AppSelect and scope selection composable for project and mission management
