### 2025-11-26 c5373b5
chore: use shared selection component
