### 2025-11-27 ed8b499
feat: first step towards new action model (#1888)
