### 2025-11-28 de39eb3
fix: fix test cases after changes to CreateTemplateDto
