### 2025-11-28 954364c
chore: use non-root user in frontent image
