+++ frontend/create_build_info.sh
@@ -1,24 +1,30 @@
 #!/bin/bash
 
-# This script extracts the build information from the frontend
+# Stop on errors
+set -e
+GIT_DIR="./.git"
 
 timestamp=$(date "+%a %b %d %Y %H:%M:%S GMT%z (%Z)")
-
 version=$(node -pe "require('./package.json').version")
 
-# check if .git directory exists and contains a HEAD file
-if [ -f ".git/HEAD" ]; then
-  git_commit=$(cat .git/$(cat .git/HEAD | cut -d' ' -f2))
-  git_branch=$(cat .git/HEAD | cut -d'/' -f3)
-  git_commit="${git_commit:0:8}"
+# Initialize default values
+git_commit="unknown"
+git_branch="unknown"
 
-else # set default values
-  git_commit="unknown"
-  git_branch="unknown"
+# Check if the required .git metadata exists
+if [ -f "$GIT_DIR/HEAD" ]; then
+     git_commit=$(git --work-tree=. --git-dir="$GIT_DIR" rev-parse --short HEAD 2>/dev/null || echo "unknown")
+
+    head_content=$(cat "$GIT_DIR/HEAD")
+
+    if [[ "$head_content" =~ "ref: refs/heads/" ]]; then
+        git_branch=$(echo "$head_content" | cut -d'/' -f3)
+    else
+        git_branch="DETACHED-$git_commit"
+    fi
 fi
 
-
-
+# Construct TypeScript content
 build_info="
 const build = {
     version: \"${version}\",
