### 2025-11-28 285e3a7
chore: fix linting and type errors
