+++ common/api/types/actions/action-templates.dto.ts
@@ -1,10 +1,10 @@
-import { ActionTemplateDto } from '@common/api/types/actions/action-template.dto';
-import { Paginated } from '@common/api/types/pagination';
-import { IsSkip } from '@common/validation/skip-validation';
-import { IsTake } from '@common/validation/take-validation';
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { IsNumber, ValidateNested } from 'class-validator';
+import { ActionTemplateDto } from '../../../api/types/actions/action-template.dto';
+import { Paginated } from '../../../api/types/pagination';
+import { IsSkip } from '../../../validation/skip-validation';
+import { IsTake } from '../../../validation/take-validation';
 
 export class ActionTemplatesDto implements Paginated<ActionTemplateDto> {
     @ApiProperty()
