### 2025-11-28 02e7363
chore: fix remaining lint issues
