### 2025-11-28 fdb9c9b
feat: make all columns sortable in actions view
