### 2025-11-28 791900c
feat: further improve action UX
