### 2025-11-28 c4b0404
fix: deletion of action
