### 2025-11-28 b115bec
fix: running action query on dashboard
