### 2025-11-28 ed5eaec
fix: adapt CLI to API changes
