+++ examples/kleinkram-actions/python-template/README.md
@@ -1,0 +1,15 @@
+# Python Template Action
+
+This is a template for creating custom actions using Python.
+
+## Description
+
+This action serves as a starting point for developing your own data processing logic. It demonstrates how to:
+- Access the Mission UUID from environment variables.
+- Download data using the `kleinkram` Python SDK.
+- Process files in the `/data` directory.
+- Save results to the `/out` directory.
+
+## Usage
+
+Use this template to build your own actions. Modify `main.py` to implement your specific logic.

[…]
