+++ examples/kleinkram-actions/validate-data/entrypoint.sh
@@ -2,6 +2,7 @@
 set -e
 
 echo "Validating data for mission $KLEINKRAM_MISSION_UUID"
+klein --version
 
 # Authenticate
 klein login --key "$KLEINKRAM_API_KEY"
