### 2025-11-29 08c21fa
feat: improve action UX
