+++ queueConsumer/src/artifactUpload/upload_artifacts.py
@@ -16,20 +16,23 @@ def upload_to_minio(file_path, bucket_name, object_name):
     # Initialize MinIO client
     # Ensure these ENVs are passed to the container
     client = Minio(
-        os.getenv("MINIO_ENDPOINT", "minio:9000"),
+        endpoint=os.getenv("MINIO_ENDPOINT", "minio:9000"),
         access_key=os.getenv("MINIO_ACCESS_KEY"),
         secret_key=os.getenv("MINIO_SECRET_KEY"),
         secure=False,  # Set to True if using HTTPS inside the network
     )
 
     # Check if bucket exists
-    if not client.bucket_exists(bucket_name):
+    if not client.bucket_exists(bucket_name=bucket_name):
         print(f"Bucket {bucket_name} does not exist. Please check configuration.")
         return
 
     # Upload
     client.fput_object(
-        bucket_name, object_name, file_path, content_type="application/gzip"
+        bucket_name=bucket_name,
+        object_name=object_name,
+        file_path=file_path,
+        content_type="application/gzip",
     )
     print(f"Successfully uploaded {object_name} to {bucket_name}")
 
