+++ docs/usage/files/files.md
@@ -2,14 +2,14 @@
 
 Kleinkram supports a variety of robotics data formats. All other files should be stored in a different Cloud Storage Solution like Google Drive.
 
-| Extension | Description |
-| :--- | :--- |
-| `.mcap` | Standard container format for multimodal data (ROS2, Foxglove). |
-| `.bag` | Legacy ROS1 bag file. Automatically converted to MCAP on upload. |
-| `.db3` | ROS2 bag file (SQLite3). |
-| `.svo2` | ZED camera recording (v2) |
-| `.tum` | TUM dataset format |
-| `.yaml` / `.yml` | YAML configuration files |
+| Extension        | Description                                                      |
+| :--------------- | :--------------------------------------------------------------- |
+| `.mcap`          | Standard container format for multimodal data (ROS2, Foxglove).  |
+| `.bag`           | Legacy ROS1 bag file. Automatically converted to MCAP on upload. |
+| `.db3`           | ROS2 bag file (SQLite3).                                         |
+| `.svo2`          | ZED camera recording (v2)                                        |
+| `.tum`           | TUM dataset format                                               |
+| `.yaml` / `.yml` | YAML configuration files                                         |
 
 ::: tip Requesting New File Types
 If you need support for additional file types, please [open a new issue](https://github.com/leggedrobotics/kleinkram/issues/new) on GitHub.
@@ -25,20 +25,21 @@ Native translation of `.bag` files to `.mcap` (using `ros1msg` encoding) is supp
 
 In the Kleinkram UI the filestates are displayed as Icons with a short explanation on hover. Here a more indepth explanation.
 
-| Icon                                                                                          | Name             | Description                                                                                                                                    |
-| --------------------------------------------------------------------------------------------- | ---------------- | ---------------------------------------------------------------------------------------------------------------------------------------------- |
-| <span class="material-symbols-outlined" style="color: green">check_circle</span>              | OK               | Everything is OK                                                                                                                               |
-| <span class="material-symbols-outlined" style="color: orange">arrow_upload_progress</span>    | Uploading        | The File is being uploaded to the File Storage. Should switch to OK when completed. Will switch to ERROR after 12h in _Uploading_.             |
-| <span class="material-symbols-outlined" style="color: red">conversion_path_off</span>         | Conversion Error | The File was uploaded, the MD5 was correct but it is not a valid bag / mcap. Check locally if it is valid. If it is, contact an Admin!         |
-| <span class="material-symbols-outlined" style="color: red">error</span>                       | Error            | Something went wrong. Contact an Admin to find out more.                                                                                       |
-| <span class="material-symbols-outlined" style="color: red">sentiment_very_dissatisfied</span> | Corrupted        | The File was uploaded but its MD5 hash doesn't match. Delete it and upload it again. Contact an Admin.                                         |
+| Icon                                                                                          | Name             | Description                                                                                                                            |
+| --------------------------------------------------------------------------------------------- | ---------------- | -------------------------------------------------------------------------------------------------------------------------------------- |
+| <span class="material-symbols-outlined" style="color: green">check_circle</span>              | OK               | Everything is OK                                                                                                                       |
+| <span class="material-symbols-outlined" style="color: orange">arrow_upload_progress</span>    | Uploading        | The File is being uploaded to the File Storage. Should switch to OK when completed. Will switch to ERROR after 12h in _Uploading_.     |
+| <span class="material-symbols-outlined" style="color: red">conversion_path_off</span>         | Conversion Error | The File was uploaded, the MD5 was correct but it is not a valid bag / mcap. Check locally if it is valid. If it is, contact an Admin! |
+| <span class="material-symbols-outlined" style="color: red">error</span>                       | Error            | Something went wrong. Contact an Admin to find out more.                                                                               |
[…]
