### 2025-11-29 03d56f1
feat: use separate endpoint for action logs (#1923)
