+++ backend/src/serialization/action.ts
@@ -2,7 +2,6 @@ import { ActionWorkerDto } from '@common/api/types/action-workers.dto';
 import { ActionDto } from '@common/api/types/actions/action.dto';
 import { AuditLogDto } from '@common/api/types/actions/audit-log.dto';
 import { DockerImageDto } from '@common/api/types/actions/docker-image.dto';
-import { LogsDto } from '@common/api/types/actions/logs.dto';
 import ActionEntity from '@common/entities/action/action.entity';
 import { actionTemplateEntityToDto } from './action-template';
 import { missionEntityToDto } from './index';
@@ -28,7 +27,7 @@ export const actionEntityToDto = (action: ActionEntity): ActionDto => {
         createdAt: action.createdAt,
         creator: userEntityToDto(action.creator),
         image: (action.image as DockerImageDto) ?? { repoDigests: [] },
-        logs: (action.logs as unknown as LogsDto[]) ?? [],
+
         mission: missionEntityToDto(action.mission),
         state: action.state,
         stateCause: action.state_cause ?? '',
