+++ common/entities/action/action.entity.ts
@@ -60,7 +60,7 @@ export default class ActionEntity extends BaseEntity {
     })
     mission?: MissionEntity;
 
-    @Column({ type: 'json', nullable: true })
+    @Column({ type: 'json', nullable: true, select: false })
     logs?: ContainerLog[];
 
     @Column({ type: 'json', nullable: true, default: [] })
