### 2025-11-29 d5b7720
fix: improve file hash missmatch error
