### 2025-11-29 7be8f88
style: reformat file using black
