### 2025-11-29 363cc80
feat: blacklist older CLI versions
