### 2025-11-29 1d6a36d
feat: secure revisions endpoint
