### 2025-11-29 0c4a471
feat: robust file events for actions
