+++ frontend/src/components/inspect-file/file-history/types.ts
@@ -1,0 +1,7 @@
+import { FileEventDto } from '@api/types/file/file-event.dto';
+
+export interface GroupedFileEvent extends FileEventDto {
+    count: number;
+    events: FileEventDto[];
+    isUploadGroup?: boolean;
+}

[…]
