### 2025-11-29 e9d2190
feat: hide GitHub and Google oAuth if not configured
