+++ frontend/src/services/auth.ts
@@ -1,4 +1,5 @@
 import { CurrentAPIUserDto } from '@api/types/user/current-api-user.dto';
+import { AvailableProvidersDto } from '@api/types/available-providers.dto';
 import axios from 'src/api/axios';
 import { getMe } from 'src/services/queries/user';
 import ENV from '../environment';
@@ -53,4 +54,9 @@ export function logout() {
 
 export const login = (provider: string): void => {
     globalThis.location.href = `${ENV.ENDPOINT}/auth/${provider}`;
+};
+
+export const getAvailableProviders = async (): Promise<AvailableProvidersDto> => {
+    const response = await axios.get<AvailableProvidersDto>('/auth/available-providers');
+    return response.data;
 };
