### 2025-11-29 f9684d8
feat: auto login using fake oauth for dev setup
