+++ cli/kleinkram/auth.py
@@ -80,8 +80,88 @@ def _browser_auth(*, url: str) -> None:
     print(f"Authentication complete. Tokens saved to {CONFIG_PATH}.")
 
 
+def _direct_oauth_auth(*, endpoint: str, provider: str, user: str) -> None:
+    """
+    Directly authenticate with fake OAuth by programmatically following the OAuth flow.
+    This bypasses the browser entirely for automated testing.
+    """
+    import requests
+    
+    print(f"Authenticating as user {user} with {provider}...")
+    
+    try:
+        # Step 1: Get the authorization code from fake OAuth
+        # The fake OAuth server will auto-redirect when user parameter is provided
+        fake_oauth_url = f"http://localhost:8004/oauth/authorize"
+        callback_url = f"{endpoint}/auth/{provider}/callback"
+        
+        params = {
+            "client_id": "some-random-string-it-does-not-matter",
+            "redirect_uri": callback_url,
+            "response_type": "code",
+            "state": "cli-direct",
+            "user": user,
+        }
+        
+        # Make request to fake OAuth - it will redirect with the auth code
+        response = requests.get(fake_oauth_url, params=params, allow_redirects=False)
+        
+        if response.status_code not in [301, 302, 303, 307, 308]:
+            raise RuntimeError(f"Expected redirect from OAuth provider, got {response.status_code}")
+        
+        # Extract the redirect location
+        location = response.headers.get("Location")
+        if not location:
+            raise RuntimeError("No redirect location from OAuth provider")
+        
+        # Parse the callback URL to extract the auth code
+        parsed = urllib.parse.urlparse(location)
+        query_params = urllib.parse.parse_qs(parsed.query)
+        
+        if "code" not in query_params:
+            raise RuntimeError(f"No authorization code in redirect: {location}")
+        
+        auth_code = query_params["code"][0]
+        state = query_params.get("state", [None])[0]
+        
+        print(f"Received authorization code, exchanging for tokens...")
+        
+        # Step 2: Exchange the code for tokens by calling the backend callback
+        # Use a session to preserve cookies
+        session = requests.Session()
+        callback_params = {"code": auth_code}
+        if state:
+            callback_params["state"] = state
+        
+        callback_response = session.get(callback_url, params=callback_params, allow_redirects=False)
+        
+        # The backend should set cookies and redirect
+        if callback_response.status_code not in [301, 302, 303, 307, 308]:
+            raise RuntimeError(f"Expected redirect from callback, got {callback_response.status_code}")
+        
+        # Extract tokens from cookies
+        auth_token = session.cookies.get("authtoken")
+        refresh_token = session.cookies.get("refreshtoken")
+        
+        if not auth_token or not refresh_token:
+            raise RuntimeError("Failed to get tokens from callback response")
+        
+        # Save tokens
+        config = get_config()
+        config.credentials = Credentials(
+            auth_token=auth_token,
+            refresh_token=refresh_token
+        )
+        save_config(config)
+        print(f"Authentication complete. Tokens saved to {CONFIG_PATH}.")
+        
+    except requests.RequestException as e:
+        raise RuntimeError(f"OAuth flow failed: {e}")
+
+
+
 def login_flow(
-    *, oAuthProvider: str, key: Optional[str] = None, headless: bool = False
+    *, oAuthProvider: str, key: Optional[str] = None, headless: bool = False, user: Optional[str] = None
 ) -> None:
     config = get_config()
     # use cli key login
@@ -90,7 +170,21 @@ def login_flow(
         save_config(config)
         return
 
[…]
