### 2025-11-30 536979d
feat: add workflow for CLI testing
