+++ .github/workflows/cli-tests.yml
@@ -1,0 +1,159 @@
+name: CLI Integration Tests
+
+on:
+    pull_request:
+        types: [opened, synchronize, reopened, edited]
+        branches:
+            - main
+            - dev
+    push:
+        branches:
+            - dev
+
+jobs:
+    test_CLI:
+        runs-on: [self-hosted, linux]
+
+        permissions:
+            contents: read
+
+        steps:
+            - uses: actions/checkout@v6
+
+            - name: Set up Python
+              uses: actions/setup-python@v5
+              with:
+                  python-version: '3.10'
+
+            - name: Build Docker stack
+              run: docker compose -f docker-compose.testing.yml build
+
+            - name: Clear previous volumes
+              run: docker compose -f docker-compose.testing.yml down -v
+
+            - name: Start fake-oauth service
+              run: |
+                  # Start fake-oauth separately since it's not in testing compose
+                  docker compose up -d fake-oauth
+
+            - name: Launch Docker stack
+              run: |
+                  docker compose -f docker-compose.testing.yml up -d
+                  echo "Docker stack started"
+
+            - name: Wait for backend to be ready
+              run: |
+                  SECONDS=0
+                  TIMEOUT=180
+                  echo "Waiting for backend to be ready..."
+                  
+                  while ! curl -s http://localhost:3000/auth/available-providers > /dev/null; do
+                    if [ $SECONDS -ge $TIMEOUT ]; then
+                        echo "Timeout reached: Backend not ready"
+                        docker compose -f docker-compose.testing.yml logs api-server
+                        docker compose -f docker-compose.testing.yml down
+                        exit 1
+                    fi
+                    sleep 2
+                    echo "Waiting for backend... ($SECONDS seconds elapsed)"
+                  done
+                  
+                  echo "Backend is ready!"
+
+            - name: Generate mock test data
+              run: |
+                  # Create test data directory if it doesn't exist
+                  mkdir -p cli/data/testing
+                  
+                  # Generate bag files with specific sizes to match cli/data/testing
+                  # Using dd to create files with random content (or zeros) of specific sizes
+                  
+                  # 10 KB
+                  dd if=/dev/zero of=cli/data/testing/10_KB.bag bs=1024 count=10
+                  
+                  # 50 KB
+                  dd if=/dev/zero of=cli/data/testing/50_KB.bag bs=1024 count=50
+                  
+                  # 1 MB
+                  dd if=/dev/zero of=cli/data/testing/1_MB.bag bs=1024 count=1024
+                  
+                  # 17 MB
+                  dd if=/dev/zero of=cli/data/testing/17_MB.bag bs=1M count=17
+                  
+                  # 125 MB
+                  dd if=/dev/zero of=cli/data/testing/125_MB.bag bs=1M count=125
+                  
+                  echo "Mock test data generated"
+
+            - name: Install CLI
+              working-directory: ./cli
+              run: |
+                  # Create virtual environment
+                  python -m venv .venv
+                  source .venv/bin/activate
+                  
+                  # Install CLI in development mode
+                  pip install -e . -r requirements.txt
+                  
+                  # Verify installation
+                  klein --version
+                  
+                  echo "CLI installed successfully"
+
+            - name: Configure CLI endpoint
+              working-directory: ./cli
+              run: |
+                  source .venv/bin/activate
+                  
+                  # Set endpoint to local
+                  klein endpoint local
+                  
+                  echo "CLI endpoint configured to local"
+
+            - name: Authenticate with fake OAuth (User 1 - Admin)
+              working-directory: ./cli
+              run: |
+                  source .venv/bin/activate
+                  
+                  # Authenticate as admin user using auto-select
+                  klein login --oauth-provider fake-oauth --user 1
+                  
+                  echo "Authenticated as admin user"
+
[…]
