### 2025-12-05 #8636 @boisgera
Hi @wp99cp,

I am about to document a first batch of points that I'd like to see addressed, in the "First contact" category. I am likely to document another "Usage" batch later (unless @c-joly ends up documenting the same points).
