### 2025-12-05 leggedrobotics/kleinkram#1977 @boisgera
First Contact (README)
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636

A few README change suggestions to (hopefully) facilitate access to information for new users

- [ ] `About`

  The link displayed in the `About` section points to the instance of Kleinkram hosted by RSL: <https://datasets.leggedrobotics.com/>

  In my opinion, this is currently confusing; by default, I would expect a link to the main project website here: <https://docs.datasets.leggedrobotics.com/>

  Pointing to the Kleinkram RSL instance could be a sensible alternative if the link to the main project was prominent in the application (typically as an extra top-level tab and/or maybe as a link in the Kleinkram logo, not as a small "Support" icon which links to a sub-section of the main web page).

- [ ] "Cloud" or Self-Hosted?

  From the README alone, it is not obvious to me that:

    - RSL hosts a Kleinkram instance at <https://datasets.leggedrobotics.com/> and

    - everyone can self-host their own instance.

- [ ] RSL Kleinkram instance status

  From the README alone, the status of the RSL instance for a random potential Kleinkram user is not clear. Is it "a cloud offering" (I can use it instead of deploying my own instance), a "private RSL instance" (I have no access), a "quick demo" to see the Kleinkram interface and get a feel for it?

  What I can discover by exploration is that: AFAICT, I have access to the instance via Google ID or GitHub (if I am not an RSL member) but only to my own workspace, which is empty and I cannot create anything. This means that the "cloud" option is out and the "quick demo" interpretation partially works.

  Please explain what I can expect from this instance! A quick, very restricted demo is fine as long as it's explained
  (even if a read-only workspace with pre-existing examples would be better than an empty one) and it's clear that I'll need to deploy
  my own Kleinkram instance to have a workable system.

- [ ] About `Getting Started` section and `Documentation`.

  I would rather have:

  - a more prominent top-level documentation link (replicated in `About`).

  - *some* of the `Getting Started` content inlined in the `README` instead of just links (`Installation` is already doing that, which is good).

- [ ] Meaning of "Developer"

  I have a bit of a hard time with the "user" vs "developer" terminology. "Kleinkram user" I think everyone can understand; but "developer" could mean anything from "developer of the Kleinkram project" (aka "contributor")  to "administrator (of a self-hosted Kleinkram instance)" to "developer of custom Kleinkram actions" (which would be sort of a "devops" role?). So at this stage I have to look at the "developer" docs to try to infer what that means.

- [ ] `Installation`.

  I can't remember off the top of my head if the default instructions yield workspaces with some mock content or empty ones, and I believe that the default configuration won't yield any write access. It could be good enough for a "Quickstart", to make sure that Kleinkram is working correctly before reading the docs, but then the limitations should be clear. And/or you could go quickly through the configuration steps that  are required to make a writable Kleinkram instance (something I struggled with).
