### 2025-12-05 leggedrobotics/kleinkram#1978 @boisgera
API Service Down (self-hosted kleinkram)
JOSS Context: https://github.com/openjournals/joss-reviews/issues/8636

Based on a fresh build of #cedbce084c677badb1d8fac2027f0cd2ae6663c2.

Afaict the API service is not functionning; attempting to log with GMail (http://localhost:3000/auth/google) yields in Firefox

```
The connection was reset

The connection to the server was reset while the page was loading.

    The site could be temporarily unavailable or too busy. Try again in a few moments.
    If you are unable to load any pages, check your computer’s network connection.
    If your computer or network is protected by a firewall or proxy, make sure that Firefox is permitted to access the web.
```

Before that I could notice the following warning in the frontend page:
