### 2025-12-05 #8636 @boisgera
Meta-issue:

  - https://github.com/leggedrobotics/kleinkram/issues/1977
  - https://github.com/leggedrobotics/kleinkram/issues/1978
  - ~~🚧 check for the many issues addressed by the upcoming 0.56 release~~
  - ~~🚧 more to come when I can log into kleinkram again~~
    **Update.** ~~blocked by https://github.com/leggedrobotics/kleinkram/issues/2045~~
    ~~**Update.** Can't get to the login page again, seems to be non-deterministic? (need some extra time on this)~~
    **Update.** OK, change of authen options implemented in local kleinkram server.
