### 2025-12-07 0cd1c9c
feat: improve docs for first contact with local setup (#1977)
