### 2025-12-07 6f96cf1
fix: unresponsive dashboard (#1733)
