### 2025-12-30 leggedrobotics/kleinkram#2045 @boisgera
Backend Unavailable
JOSS context: https://github.com/openjournals/joss-reviews/issues/8636
cc: @c-joly

The front-end displays the following error when I start locally the kleinkram server (version 0.56.0):

Additional context:

  - A `qemu-system-x86` process consumes a lot of CPU for quite a long time (transitionally). Is it related?

  - Some of the log :

  ```
  kleinkram-loki               | level=error ts=2025-12-30T10:00:14.706543839Z caller=scheduler_processor.go:110 component=querier msg="error processing requests from scheduler" err="rpc error: code = Unavailable desc = keepalive ping failed to receive ACK within timeout" addr=127.0.0.1:9095
  kleinkram-tempo              | level=error ts=2025-12-30T10:00:14.750065337Z caller=frontend_processor.go:84 msg="error processing requests" address=127.0.0.1:9095 err="rpc error: code = Unavailable desc = keepalive ping failed to receive ACK within timeout"
  kleinkram-tempo              | level=error ts=2025-12-30T10:00:14.758782515Z caller=frontend_processor.go:84 msg="error processing requests" address=127.0.0.1:9095 err="rpc error: code = Unavailable desc = keepalive ping failed to receive ACK within timeout"
  kleinkram-queue-consumer     | [ERROR]: Unable to connect to the database. Retrying (7)...

  v View in Docker Desktop   o View Config   w Enable Watchkleinkram-tempo              | time=2025-12-30T10:00:14.762Z level=WARN source=/home/runner/work/tempo/tempo/vendor/github.com/prometheus/prometheus/storage/remote/queue_manager.go:2068 msg="Failed to send batch, retrying" tenant=single-tenant component=remote remote_name=87de1c url=http://prometheus:9090/api/v1/write err="Post \"http://prometheus:9090/api/v1/write\": context deadline exceeded"
  kleinkram-loki               | level=error ts=2025-12-30T10:00:14.707711168Z caller=scheduler_processor.go:110 component=querier msg="error processing requests from scheduler" err="rpc error: code = Unavailable desc = keepalive ping failed to receive ACK within timeout" addr=127.0.0.1:9095
  kleinkram-loki               | level=error ts=2025-12-30T10:00:14.709929811Z caller=scheduler_processor.go:110 component=querier msg="error processing requests from scheduler" err="rpc error: code = Unavailable desc = keepalive ping failed to receive ACK within timeout" addr=127.0.0.1:9095
  kleinkram-loki               | level=error ts=2025-12-30T10:00:14.705547373Z caller=scheduler_processor.go:110 component=querier msg="error processing requests from scheduler" err="rpc error: code = Unavailable desc = keepalive ping failed to receive ACK within timeout" addr=127.0.0.1:9095
  kleinkram-tempo              | level=error ts=2025-12-30T10:00:14.776138951Z caller=frontend_processor.go:84 msg="error processing requests" address=127.0.0.1:9095 err="rpc error: code = Unavailable desc = keepalive ping failed to receive ACK within timeout"
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:14.783Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-grafana            | t=2025-12-30T10:00:14.796585301Z level=error caller=logger.go:234 time=2025-12-30T10:00:14.790284501Z msg="cleaning up inactive secure values" error="fetching inactive secure values that need to be cleaned up: acquiring leases for inactive secure values: leasing inactive secure values: context deadline exceeded"
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:15.784Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:15.801Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:15.818Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-api-server         |
  kleinkram-api-server         | > @kleinkram/backend-common@0.56.0 seed:run /app/packages/backend-common
  kleinkram-api-server         | > ts-node --project tsconfig.json -r tsconfig-paths/register $(node -p "require.resolve('typeorm-seeding/dist/cli.js')") seed --configName src/ormconfig.ts
  kleinkram-api-server         |
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:16.819Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:17.278Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:17.289Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-queue-consumer     | [ERROR]: Unable to connect to the database. Retrying (8)...
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:18.290Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-api-server         | 🌱  TypeORM Seeding v1.6.1
  kleinkram-api-server         | - Loading ormconfig
  kleinkram-tempo              | time=2025-12-30T10:00:19.657Z level=INFO source=/home/runner/work/tempo/tempo/vendor/github.com/prometheus/prometheus/storage/remote/queue_manager.go:1096 msg="Remote storage resharding" tenant=single-tenant component=remote remote_name=87de1c url=http://prometheus:9090/api/v1/write from=1 to=28
  kleinkram-queue-consumer     | [ERROR]: Unable to connect to the database. Retrying (9)...
  kleinkram-queue-consumer     | [ERROR]: getaddrinfo ENOTFOUND database
  kleinkram-api-server         | ✔ ORM Config loaded
  kleinkram-api-server         | - Import Factories
  kleinkram-api-server         | webpack 5.103.0 compiled successfully in 151938 ms
  kleinkram-api-server         | ✔ Factories are imported
  kleinkram-api-server         | - Importing Seeders
  kleinkram-api-server         | ✔ Seeders are imported
  kleinkram-api-server         | - Connecting to the database
  kleinkram-api-server         | ✖ Database connection failed! Check your typeORM config file.
  kleinkram-api-server         | Error: getaddrinfo ENOTFOUND database
  kleinkram-api-server         |     at GetAddrInfoReqWrap.onlookupall [as oncomplete] (node:dns:122:26) {
  kleinkram-api-server         |   errno: -3008,
  kleinkram-api-server         |   code: 'ENOTFOUND',
  kleinkram-api-server         |   syscall: 'getaddrinfo',
  kleinkram-api-server         |   hostname: 'database'
  kleinkram-api-server         | }
  kleinkram-api-server         |  ELIFECYCLE  Command failed with exit code 1.
  kleinkram-api-server         | ----------------------------------------
  kleinkram-api-server         | Backend API ready on http://localhost:3000
  kleinkram-api-server         | ----------------------------------------
  kleinkram-api-server         | [INFO]: Starting Nest application...
  kleinkram-api-server         | [DEBUG]: DBDumper service created
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (1)...
  kleinkram-api-server         | Found controller: ActionsController
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:27.276Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:27.291Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:28.292Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (2)...
  kleinkram-api-server         | Found controller: AccessController
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (3)...
  kleinkram-api-server         | Found controller: AuthController
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (4)...
  kleinkram-api-server         | Found controller: CategoryController
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:37.276Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:37.289Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:38.290Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (5)...
  kleinkram-api-server         | Found controller: FileController
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (6)...
  kleinkram-api-server         | Found controller: HealthController
  kleinkram-api-server         | Found controller: FoxgloveController
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (7)...
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:47.383Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:47.396Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-api-server         | Found controller: MissionController
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (8)...
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:48.396Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-api-server         | Found controller: ProjectController
  kleinkram-api-server         | Found controller: OldProjectController
  kleinkram-api-server         | [ERROR]: Unable to connect to the database. Retrying (9)...
  kleinkram-api-server         | [ERROR]: getaddrinfo ENOTFOUND database
  kleinkram-api-server         | Found controller: TagController
  kleinkram-api-server         | Found controller: TemplatesController
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:57.383Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:57.399Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:00:58.399Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-api-server         | Found controller: TopicController
  kleinkram-tempo              | time=2025-12-30T10:00:59.765Z level=WARN source=/home/runner/work/tempo/tempo/vendor/github.com/prometheus/prometheus/storage/remote/queue_manager.go:1117 msg="Skipping resharding, last successful send was beyond threshold" tenant=single-tenant component=remote remote_name=87de1c url=http://prometheus:9090/api/v1/write lastSendTimestamp=1767088844 minSendTimestamp=1767088849
  kleinkram-api-server         | Found controller: UserController
  kleinkram-api-server         | Found controller: WorkerController
  kleinkram-api-server         | Initializing NestJS TestingModule with 15 controllers...
  kleinkram-api-server         | swagger.json generated successfully.
  kleinkram-api-server         | Endpoints saved to /app/docs/development/api/generated/__generated__endpoints.json
  kleinkram-api-server         | Generating api-modules.md...
  kleinkram-api-server         | api-modules.md, api-modules.json and individual module files generated successfully.
  kleinkram-api-server         | npm notice
  kleinkram-api-server         | npm notice New major version of npm available! 10.9.4 -> 11.7.0
  kleinkram-api-server         | npm notice Changelog: https://github.com/npm/cli/releases/tag/v11.7.0
  kleinkram-api-server         | npm notice To update run: npm install -g npm@11.7.0
  kleinkram-api-server         | npm notice
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:07.382Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:07.396Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:08.398Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:17.382Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:17.397Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:18.397Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:27.382Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:27.397Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:28.399Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:37.382Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:37.402Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:38.403Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:47.382Z level=INFO source=server.go:73 msg="Established new database connection" fingerprint=database:5432
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:47.396Z level=ERROR source=collector.go:178 msg="Error opening connection to database" err="error querying postgresql version: dial tcp: lookup database on 127.0.0.11:53: no such host"
  kleinkram-postgres-exporter  | time=2025-12-30T10:01:48.396Z level=ERROR source=postgres_exporter.go:684 msg="error scraping dsn" err="Error opening connection to database (postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable): dial tcp: lookup database on 127.0.0.11:53: no such host" dsn="postgresql://dbuser:PASSWORD_REMOVED@database:5432/dbname?sslmode=disable"
  kleinkram-tempo              | time=2025-12-30T10:01:49.766Z level=INFO source=/home/runner/work/tempo/tempo/vendor/github.com/prometheus/prometheus/storage/remote/queue_manager.go:1096 msg="Remote storage resharding" tenant=single-tenant component=remote remote_name=87de1c url=http://prometheus:9090/api/v1/write from=2 to=1
  ```
