### 2025-12-30 leggedrobotics/kleinkram#1977 @boisgera
Hi @wp99cp,

In my opinion, the README and documentation are much better now after the 0.56.0 release! Thanks for your work on this 🙏
