### 2025-12-30 leggedrobotics/kleinkram#2045 @wp99cp
First, make sure you have cleared any existing volumes following step 2 of the docs: https://docs.datasets.dev.leggedrobotics.com/development/try-locally

Alternatively, you can run `docker system prune --all --volumes` to clear all Docker-related data from your system. WARNING: This may delete the date of other Docker-based projects on your system.

## Advanced Debugging Steps
Seems like your PostgreSQL Docker container is not running

```
kleinkram-api-server         | ✖ Database connection failed! Check your typeORM config file.
kleinkram-api-server         | Error: getaddrinfo ENOTFOUND database
kleinkram-api-server         |     at GetAddrInfoReqWrap.onlookupall [as oncomplete] (node:dns:122:26) {
kleinkram-api-server         |   errno: -3008,
kleinkram-api-server         |   code: 'ENOTFOUND',
kleinkram-api-server         |   syscall: 'getaddrinfo',
kleinkram-api-server         |   hostname: 'database'
kleinkram-api-server         | }
```

Is any PostgreSQL container visible when running `docker ps`? Is any other app using the PostgreSQL default port? What's the output of `lsof -i :5432`.

If you run `docker compose up database` in a separate terminal, you should get something like:

```
❯ docker compose up database
[+] up 2/2
 ✔ Network kleinkram_private-network Created                                                                                                                                                0.0s
 ✔ Container kleinkram-postgres      Created                                                                                                                                                0.1s
Attaching to kleinkram-postgres
kleinkram-postgres  |
kleinkram-postgres  | PostgreSQL Database directory appears to contain a database; Skipping initialization
kleinkram-postgres  |
kleinkram-postgres  | 2025-12-30 10:23:40.185 UTC [1] LOG:  starting PostgreSQL 17.7 (Debian 17.7-3.pgdg13+1) on x86_64-pc-linux-gnu, compiled by gcc (Debian 14.2.0-19) 14.2.0, 64-bit
kleinkram-postgres  | 2025-12-30 10:23:40.185 UTC [1] LOG:  listening on IPv4 address "0.0.0.0", port 5432
kleinkram-postgres  | 2025-12-30 10:23:40.185 UTC [1] LOG:  listening on IPv6 address "::", port 5432
kleinkram-postgres  | 2025-12-30 10:23:40.190 UTC [1] LOG:  listening on Unix socket "/var/run/postgresql/.s.PGSQL.5432"
kleinkram-postgres  | 2025-12-30 10:23:40.195 UTC [28] LOG:  database system was interrupted; last known up at 2025-12-19 15:07:54 UTC
kleinkram-postgres  | 2025-12-30 10:23:40.277 UTC [28] LOG:  database system was not properly shut down; automatic recovery in progress
kleinkram-postgres  | 2025-12-30 10:23:40.280 UTC [28] LOG:  redo starts at 0/1AAD740
kleinkram-postgres  | 2025-12-30 10:23:40.282 UTC [28] LOG:  invalid record length at 0/1AEADA0: expected at least 24, got 0
kleinkram-postgres  | 2025-12-30 10:23:40.282 UTC [28] LOG:  redo done at 0/1AEAD78 system usage: CPU: user: 0.00 s, system: 0.00 s, elapsed: 0.00 s
kleinkram-postgres  | 2025-12-30 10:23:40.291 UTC [26] LOG:  checkpoint starting: end-of-recovery immediate wait
kleinkram-postgres  | 2025-12-30 10:23:40.336 UTC [26] LOG:  checkpoint complete: wrote 37 buffers (0.2%); 0 WAL file(s) added, 0 removed, 0 recycled; write=0.001 s, sync=0.031 s, total=0.048 s; sync files=13, longest=0.003 s, average=0.003 s; distance=245 kB, estimate=245 kB; lsn=0/1AEADA0, redo lsn=0/1AEADA0
kleinkram-postgres  | 2025-12-30 10:23:40.341 UTC [1] LOG:  database system is ready to accept connections
```
