### 2025-12-30 #8636 @boisgera
Extra minor issues (non-blocking for the JOSS submission):

  - https://github.com/leggedrobotics/kleinkram/pull/2046
  - https://github.com/leggedrobotics/kleinkram/pull/2047
