### 2025-12-30 leggedrobotics/kleinkram#2045 @wp99cp
Side note on `qemu-system-x86`, this is a VM process, probably as you are using Docker Desktop.
Docker Desktop creates a Linux VM to run the containers. See https://docs.docker.com/desktop/troubleshoot-and-support/faqs/linuxfaqs/#why-does-docker-desktop-for-linux-run-a-vm
