### 2026-01-09 f695359
docs: improve kleinkram action docs (#1759)
