+++ docs/development/getting-started.md
@@ -145,7 +145,7 @@ For a complete list of what is seeded, check out the seed files in `common/seeds
 :::
 
 ::: tip Actions
-More info about actions can be found in the [actions](/usage/actions/usage) section.
+More info about actions can be found in the [actions](/usage/actions/use-actions) section.
 :::
 
 ### Configuration
