+++ docs/usage/faq.md
@@ -1,7 +1,7 @@
 # Frequently Asked Questions
 
 ::: details How do I process my data?
-You can use Actions to validate, convert, and process your data. See the [Actions guide](actions/usage.md).
+You can use Actions to validate, convert, and process your data. See the [Actions guide](actions/use-actions.md).
 :::
 
 ::: details What file formats are supported?
@@ -13,7 +13,7 @@ You can use our Python SDK to script and automate your interactions. See the [Py
 :::
 
 ::: details How to write a custom action?
-See the [Developer Guide](actions/developing.md) for instructions on creating custom actions.
+See the [Developer Guide](actions/write-actions.md) for instructions on creating custom actions.
 :::
 
 ::: details How does the access control work?
