### 2026-01-22 #8636 @boisgera
cc: @c-joly

Hi @wp99cp,

My mandatory checklist is down to two unchecked items, here are the corresponding (small) issues:

 - wrt **Comunity guidelines.** Please add in the README some basic "guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support". (You already have the hard part done, which is the developer docs ; but that does not explain for example how one could upstream its contribution)

  - wrt **State of the field** . Your paper only mentions tangentially what the "competitors" would be ("cloud solutions", "Google drive"). It would be nice to say what solutions some of the big consumers of ROSbag/MCAP datasets are typically using (e.g. HuggingFace? Directly S3? other local solutions? etc) *if you have more concrete information about this*. Given the description of kleinkram characteristic that you already give in the paper, I think that the comparison would be automatic.
