### 2026-01-27 #8636 @wp99cp
cc: @c-joly

Hi @boisgera
Thanks for your feedback, once again, very helpful. I addressed the two remaining points; the corresponding commits are on the dev branch and will be included in the next release.

- wrt State of the field: I updated the `Statement of need` section in the paper to include a bit more detail of the competitors and their limitations. For now, I switched the branch to `dev` and regenerated the paper pdf using the editorialbot
- wrt Comunity guidelines: I updated the readme, see https://github.com/leggedrobotics/kleinkram/tree/dev, and added a section to the docs. Keep in mind that the `Contributing Guide` link is broken until we merge dev into main.
