### 2026-02-16 #8636 @boisgera
(with the editor hat on)

Hi @wp99cp,

@c-joly told me privately of a blocker to continue his review: https://github.com/leggedrobotics/kleinkram/issues/2101

Best,

Sébastien
