### 2026-02-18 #8636 @c-joly
Hi @wp99cp ,
Cc @boisgera

I finally succeeded to validate the installation from scratch (last week, the building process stucked for 3 hours without extra log...). Will be able to complete the review soon.

Best,

Cyril
