### 2026-02-20 #8636 @c-joly
Hi @wp99cp,
Cc @boisgera

Sorry for the delay, I was stuck in December and January with my classes and grading students. I finally finished all the tests I wanted to make. There was a huge improvement since the last version I tested in November, both for Kleinkram itself and its documentation. In particular, the way to use the application and to deal with actions is much smoother! Congratulations to the dev 👏
In my opinion, the project has reached a good state for publication; thus I validated all the checkboxes. @wp99cp: please don't forget to merge the "Contributing" section to the `main` branch (it is currently only in the `dev` branch).

Best regards,

Cyril
