### 2026-02-25 #8636 @boisgera
Hi @wp99cp,

We have entered the [post-review stage](https://joss.readthedocs.io/en/latest/editing.html#post-review) of your submission.

Could you please perform the tasks in the **Additional Author Tasks After Review is Complete** list above?

Best,

Sébastien
