### 2026-02-26 #8636 @wp99cp
@boisgera thanks for the review.

We are currently working to replace the project's no-longer-supported [Minio S3 storage](https://github.com/minio/minio) backend. As soon as this migration is complete, we will merge the dev branch, create a new release, and then be officially ready to publish the paper. We should not publish this work until that is fixed, as Minio has been marked as archived since `Feb 13, 2026`.

Sorry for another delay.
