### 2026-03-04 #8636 @wp99cp
- [x] Make a release of the software with the latest changes from the review --> that is version 0.59.0
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here --> https://doi.org/20.500.11850/794782
