### 2026-03-16 #8636 @wp99cp
@boisgera, finally we are ready!

We have successfully finished the transition from minio to SeaweedFS in Version 0.59.0. I made a new release and published it to the official ETH software archive (https://doi.org/20.500.11850/794782). With that, we have completed all the steps described in `Additional Author Tasks After Review is Complete` list.

Thanks for your patience.
Let me know if you need anything else.
