### 2026-03-27 #8636 @wp99cp
> Hi [@wp99cp](https://github.com/wp99cp),
>
> I see that your DOI is a metadata-only entry. But JOSS requires an archive of the software distribution associated to the DOI for long-term preservation (see for example [a recent Zenodo archive + DOI](https://zenodo.org/records/19046304)).
>
> * Can you add the full source distribution of the kleinkram 0.59.0 release to your entry?
> * Could you also reclassify the entry as "Software" (instead of "Working Paper")?
> * Then please align the edition/version with you release (0.59.0) and document the Right/License metadata
>
> I think that [this ETH Research collection entry](https://doi.org/10.3929/ethz-b-000524039) is a good example of what we would need.

Sorry for the delay. Give me another week to complete the data upload to our university's archive program. I will get back to you as soon as possible.
