### 2026-04-28 #8636 @boisgera
Hi @wp99cp,

Any update on your side?

(To be clear, at this stage of the process, there is not hurry on the JOSS side, I am merely checking the project state from time to time.)

Best,

SB
